+++ lib/jslint-2015-10-29.js
@@ -1,5 +1,5 @@
 // jslint.js
-// 2016-05-10
+// 2015-10-29
 // Copyright (c) 2015 Douglas Crockford  (www.JSLint.com)
 
 // Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -50,7 +50,7 @@
 //      tokens: an array of objects representing the tokens in the file.
 //      tree: the token objects arranged in a tree.
 //      warnings: an array of warning objects. A warning object can contain:
-//          name: "JSLintError"
+//          name: 'JSLintError'
 //          column: A column number in the file.
 //          line: A line number in the file.
 //          code: A warning code string.
@@ -76,54 +76,54 @@
 //      5. Check the whitespace between the tokens.
 
 // jslint can also examine JSON text. It decides that a file is JSON text if
-// the first token is "[" or "{". Processing of JSON text is much simpler than
+// the first token is '[' or '{'. Processing of JSON text is much simpler than
 // the processing of JavaScript programs. Only the first three phases are
 // required.
 
 // WARNING: JSLint will hurt your feelings.
 
 /*property
-    a, and, arity, b, bad_assignment_a, bad_directive_a, bad_get,
-    bad_module_name_a, bad_option_a, bad_property_a, bad_set, bitwise, block,
-    body, browser, c, calls, catch, charAt, charCodeAt, closer, closure, code,
-    column, concat, constant, context, couch, create, d, dead, devel,
-    directive, directives, disrupt, dot, duplicate_a, edition, ellipsis, else,
-    empty_block, es6, eval, every, expected_a_at_b_c, expected_a_b,
-    expected_a_b_from_c_d, expected_a_before_b, expected_digits_after_a,
-    expected_four_digits, expected_identifier_a, expected_line_break_a_b,
-    expected_regexp_factor_a, expected_space_a_b, expected_statements_a,
-    expected_string_a, expected_type_string_a, expression, extra, flag, for,
-    forEach, free, from, fud, fudge, function, function_in_loop, functions, g,
-    global, i, id, identifier, import, imports, inc, indexOf, infix_in, init,
-    initial, isArray, isNaN, join, json, keys, label, label_a, lbp, led,
-    length, level, line, lines, live, loop, m, margin, match, maxerr, maxlen,
-    message, misplaced_a, misplaced_directive_a, missing_browser, module,
-    multivar, naked_block, name, names, nested_comment, new, node, not_label_a,
+    a, and, arity, b, bad_assignment_a, bad_character_number_a,
+    bad_directive_a, bad_get, bad_module_name_a, bad_option_a, bad_property_a,
+    bad_set, bitwise, block, body, browser, c, calls, catch, charAt,
+    charCodeAt, closer, closure, code, column, concat, constant, context,
+    couch, create, d, dead, devel, directive, disrupt, dot, duplicate_a,
+    edition, ellipsis, else, empty_block, es6, eval, every, expected_a_at_b_c,
+    expected_a_b, expected_a_b_from_c_d, expected_a_before_b,
+    expected_digits_after_a, expected_four_digits, expected_identifier_a,
+    expected_line_break_a_b, expected_regexp_factor_a, expected_space_a_b,
+    expected_statements_a, expected_string_a, expected_type_string_a,
+    expression, extra, flag, for, forEach, free, from, fud, fudge, function,
+    function_in_loop, functions, g, global, i, id, identifier, import, imports,
+    inc, indexOf, infix_in, init, initial, isArray, isNaN, join, json, keys,
+    label, label_a, lbp, led, length, level, line, lines, live, loop, m,
+    margin, match, maxerr, maxlen, message, misplaced_a, misplaced_directive_a,
+    module, naked_block, name, names, nested_comment, new, node, not_label_a,
     nud, ok, open, option, out_of_scope_a, parameters, pop, property, push,
     qmark, quote, redefinition_a_b, replace, reserved_a, role, search,
-    signature, slice, some, sort, split, statement, stop, strict, subscript_a,
-    switch, test, this, thru, toString, todo_comment, tokens, too_long,
-    too_many, too_many_digits, tree, type, u, unclosed_comment, unclosed_mega,
+    signature, slash_equal, slice, some, sort, split, statement, stop, strict,
+    subscript_a, switch, test, this, thru, toString, todo_comment, tokens,
+    too_long, too_many, tree, type, u, unclosed_comment, unclosed_mega,
     unclosed_string, undeclared_a, unexpected_a, unexpected_a_after_b,
     unexpected_at_top_level_a, unexpected_char_a, unexpected_comment,
     unexpected_directive_a, unexpected_expression_a, unexpected_label_a,
-    unexpected_parens, unexpected_space_a_b, unexpected_statement_a,
-    unexpected_trailing_space, unexpected_typeof_a, uninitialized_a,
-    unreachable_a, unregistered_property_a, unsafe, unused_a, use_spaces, used,
-    value, var_loop, var_switch, variable, warning, warnings,
+    unexpected_parens, unexpected_space_a_b,
+    unexpected_statement_a, unexpected_trailing_space, unexpected_typeof_a,
+    uninitialized_a, unreachable_a, unregistered_property_a, unsafe, unused_a,
+    use_spaces, used, value, var_loop, var_switch, variable, warning, warnings,
     weird_condition_a, weird_expression_a, weird_loop, weird_relation_a, white,
     wrap_assignment, wrap_condition, wrap_immediate, wrap_regexp, wrap_unary,
     wrapped, writable, y
 */
 
 var jslint = (function JSLint() {
-    "use strict";
+    'use strict';
 
     function empty() {
 
 // The empty function produces a new empty object that inherits nothing. This is
 // much better than {} because confusions around accidental method names like
-// "constructor" are completely avoided.
+// 'constructor' are completely avoided.
 
         return Object.create(null);
     }
@@ -146,36 +146,35 @@ var jslint = (function JSLint() {
 
         bitwise: true,
         browser: [
-            "Audio", "clearInterval", "clearTimeout", "document", "event",
-            "FormData", "history", "Image", "localStorage", "location", "name",
-            "navigator", "Option", "screen", "sessionStorage", "setInterval",
-            "setTimeout", "Storage", "XMLHttpRequest"
+            'Audio', 'clearInterval', 'clearTimeout', 'document', 'event',
+            'FormData', 'history', 'Image', 'localStorage', 'location', 'name',
+            'navigator', 'Option', 'screen', 'sessionStorage', 'setInterval',
+            'setTimeout', 'Storage', 'XMLHttpRequest'
         ],
         couch: [
-            "emit", "getRow", "isArray", "log", "provides", "registerType",
-            "require", "send", "start", "sum", "toJSON"
+            'emit', 'getRow', 'isArray', 'log', 'provides', 'registerType',
+            'require', 'send', 'start', 'sum', 'toJSON'
         ],
         devel: [
-            "alert", "confirm", "console", "Debug", "opera", "prompt", "WSH"
+            'alert', 'confirm', 'console', 'Debug', 'opera', 'prompt', 'WSH'
         ],
         es6: [
-            "ArrayBuffer", "DataView", "Float32Array", "Float64Array",
-            "Generator", "GeneratorFunction", "Int8Array", "Int16Array",
-            "Int32Array", "Intl", "Map", "Promise", "Proxy", "Reflect",
-            "Set", "Symbol", "System", "Uint8Array", "Uint8ClampedArray",
-            "Uint16Array", "Uint32Array", "WeakMap", "WeakSet"
+            'ArrayBuffer', 'DataView', 'Float32Array', 'Float64Array',
+            'Generator', 'GeneratorFunction', 'Int8Array', 'Int16Array',
+            'Int32Array', 'Intl', 'Map', 'Promise', 'Proxy', 'Reflect',
+            'Set', 'Symbol', 'System', 'Uint8Array', 'Uint8ClampedArray',
+            'Uint16Array', 'Uint32Array', 'WeakMap', 'WeakSet'
         ],
         eval: true,
         for: true,
         fudge: true,
         maxerr: 10000,
         maxlen: 10000,
-        multivar: true,
         node: [
-            "Buffer", "clearImmediate", "clearInterval", "clearTimeout",
-            "console", "exports", "global", "module", "process", "querystring",
-            "require", "setImmediate", "setInterval", "setTimeout",
-            "__dirname", "__filename"
+            'Buffer', 'clearImmediate', 'clearInterval', 'clearTimeout',
+            'console', 'exports', 'global', 'module', 'process',
+            'require', 'setImmediate', 'setInterval', 'setTimeout',
+            '__dirname', '__filename'
         ],
         this: true,
         white: true
@@ -185,92 +184,92 @@ var jslint = (function JSLint() {
 
 // This is the set of infix operators that require a space on each side.
 
-        "!=": true,
-        "!==": true,
-        "%": true,
-        "%=": true,
-        "&": true,
-        "&=": true,
-        "&&": true,
-        "*": true,
-        "*=": true,
-        "+=": true,
-        "-=": true,
-        "/": true,
-        "/=": true,
-        "<": true,
-        "<=": true,
-        "<<": true,
-        "<<=": true,
-        "=": true,
-        "==": true,
-        "===": true,
-        "=>": true,
-        ">": true,
-        ">=": true,
-        ">>": true,
-        ">>=": true,
-        ">>>": true,
-        ">>>=": true,
-        "^": true,
-        "^=": true,
-        "|": true,
-        "|=": true,
-        "||": true
+        '!=': true,
+        '!==': true,
+        '%': true,
+        '%=': true,
+        '&': true,
+        '&=': true,
+        '&&': true,
+        '*': true,
+        '*=': true,
+        '+=': true,
+        '-=': true,
+        '/': true,
+        '/=': true,
+        '<': true,
+        '<=': true,
+        '<<': true,
+        '<<=': true,
+        '=': true,
+        '==': true,
+        '===': true,
+        '=>': true,
+        '>': true,
+        '>=': true,
+        '>>': true,
+        '>>=': true,
+        '>>>': true,
+        '>>>=': true,
+        '^': true,
+        '^=': true,
+        '|': true,
+        '|=': true,
+        '||': true
     };
 
     var bitwiseop = {
 
 // These are the bitwise operators.
 
-        "~": true,
-        "^": true,
-        "^=": true,
-        "&": true,
-        "&=": true,
-        "|": true,
-        "|=": true,
-        "<<": true,
-        "<<=": true,
-        ">>": true,
-        ">>=": true,
-        ">>>": true,
-        ">>>=": true
+        '~': true,
+        '^': true,
+        '^=': true,
+        '&': true,
+        '&=': true,
+        '|': true,
+        '|=': true,
+        '<<': true,
+        '<<=': true,
+        '>>': true,
+        '>>=': true,
+        '>>>': true,
+        '>>>=': true
     };
 
     var opener = {
 
 // The open and close pairs.
 
-        "(": ")",       // paren
-        "[": "]",       // bracket
-        "{": "}",       // brace
-        "${": "}"       // mega
+        '(': ')',       // paren
+        '[': ']',       // bracket
+        '{': '}',       // brace
+        '${': '}'       // mega
     };
 
     var relationop = {
 
 // The relational operators.
 
-        "!=": true,
-        "!==": true,
-        "==": true,
-        "===": true,
-        "<": true,
-        "<=": true,
-        ">": true,
-        ">=": true
+        '!=': true,
+        '!==': true,
+        '==': true,
+        '===': true,
+        '<': true,
+        '<=': true,
+        '>': true,
+        '>=': true
     };
 
     var standard = [
 
 // These are the globals that are provided by the ES5 language standard.
 
-        "Array", "Boolean", "Date", "decodeURI", "decodeURIComponent",
-        "encodeURI", "encodeURIComponent", "Error", "EvalError", "isFinite",
-        "isNaN", "JSON", "Math", "Number", "Object", "parseInt", "parseFloat",
-        "RangeError", "ReferenceError", "RegExp", "String", "SyntaxError",
-        "TypeError", "URIError"
+        'Array', 'Boolean', 'Date', 'decodeURI', 'decodeURIComponent',
+        'encodeURI', 'encodeURIComponent', 'Error', 'EvalError', 'Function',
+        'isFinite', 'isNaN', 'JSON', 'Math', 'Number', 'Object', 'parseInt',
+        'parseFloat', 'RangeError', 'ReferenceError', 'RegExp', 'String',
+        'SyntaxError', 'TypeError', 'URIError'
     ];
 
     var bundle = {
@@ -284,6 +283,7 @@ var jslint = (function JSLint() {
 
         and: "The '&&' subexpression should be wrapped in parens.",
         bad_assignment_a: "Bad assignment to '{a}'.",
+        bad_character_number_a: "Bad character code: '{a}'",
         bad_directive_a: "Bad directive '{a}'.",
         bad_get: "A get function takes no parameters.",
         bad_module_name_a: "Bad module name '{a}'.",
@@ -292,7 +292,7 @@ var jslint = (function JSLint() {
         bad_set: "A set function takes one parameter.",
         duplicate_a: "Duplicate '{a}'.",
         empty_block: "Empty block.",
-        es6: "Unexpected ES6 feature '{a}'.",
+        es6: "Unexpected ES6 feature.",
         expected_a_at_b_c: "Expected '{a}' at column {b}, not column {c}.",
         expected_a_b: "Expected '{a}' and instead saw '{b}'.",
         expected_a_b_from_c_d: "Expected '{a}' to match '{b}' from line {c} and instead saw '{d}'.",
@@ -312,18 +312,17 @@ var jslint = (function JSLint() {
         label_a: "'{a}' is a statement label.",
         misplaced_a: "Place '{a}' at the outermost level.",
         misplaced_directive_a: "Place the '/*{a}*/' directive before the first statement.",
-        missing_browser: "/*global*/ requires the Assume a browser option.",
         naked_block: "Naked block.",
         nested_comment: "Nested comment.",
         not_label_a: "'{a}' is not a label.",
         out_of_scope_a: "'{a}' is out of scope.",
         redefinition_a_b: "Redefinition of '{a}' from line {b}.",
         reserved_a: "Reserved name '{a}'.",
+        slash_equal: "A regular expression literal can be confused with '/='.",
         subscript_a: "['{a}'] is better written in dot notation.",
         todo_comment: "Unexpected TODO comment.",
         too_long: "Line too long.",
         too_many: "Too many warnings.",
-        too_many_digits: "Too many digits.",
         unclosed_comment: "Unclosed comment.",
         unclosed_mega: "Unclosed mega literal.",
         unclosed_string: "Unclosed string.",
@@ -366,46 +365,44 @@ var jslint = (function JSLint() {
 // Regular expression literals:
 
 // supplant {variables}
-    var rx_supplant = /\{([^{}]*)\}/g;
+    var rx_supplant = /\{([^{}]*)\}/g,
 // carriage return, carriage return linefeed, or linefeed
-    var rx_crlf = /\n|\r\n?/;
+        rx_crlf = /\n|\r\n?/,
 // unsafe characters that are silently deleted by one or more browsers
-    var rx_unsafe = /[\u0000-\u001f\u007f-\u009f\u00ad\u0600-\u0604\u070f\u17b4\u17b5\u200c-\u200f\u2028-\u202f\u2060-\u206f\ufeff\ufff0-\uffff]/;
+        rx_unsafe = /[\u0000-\u001f\u007f-\u009f\u00ad\u0600-\u0604\u070f\u17b4\u17b5\u200c-\u200f\u2028-\u202f\u2060-\u206f\ufeff\ufff0-\uffff]/,
 // identifier
-    var rx_identifier = /^([a-zA-Z_$][a-zA-Z0-9_$]*)$/;
-    var rx_bad_property = /^_|\$|Sync$|_$/;
+        rx_identifier = /^([a-zA-Z_$][a-zA-Z0-9_$]*)$/,
+        rx_bad_property = /^_|\$|Sync$|_$/,
 // star slash
-    var rx_star_slash = /\*\//;
+        rx_star_slash = /\*\//,
 // slash star
-    var rx_slash_star = /\/\*/;
+        rx_slash_star = /\/\*/,
 // slash star or ending slash
-    var rx_slash_star_or_slash = /\/\*|\/$/;
+        rx_slash_star_or_slash = /\/\*|\/$/,
 // uncompleted work comment
-    var rx_todo = /\b(?:todo|TO\s?DO|HACK)\b/;
+        rx_todo = /\b(?:todo|TO\s?DO|HACK)\b/,
 // tab
-    var rx_tab = /\t/g;
+        rx_tab = /\t/g,
 // directive
-    var rx_directive = /^(jslint|property|global)\s+(.*)$/;
-    var rx_directive_part = /^([a-zA-Z$_][a-zA-Z0-9$_]*)\s*(?::\s*(true|false|[0-9]+)\s*)?(?:,\s*)?(.*)$/;
+        rx_directive = /^(jslint|property|global)\s*(.*)$/,
+        rx_directive_part = /^([a-zA-Z$_][a-zA-Z0-9$_]*)\s*(?::\s*(true|false|[0-9]+)\s*)?(?:,\s*)?(.*)$/,
 // token (sorry it is so long)
-    var rx_token = /^((\s+)|([a-zA-Z_$][a-zA-Z0-9_$]*)|[(){}\[\]\?,:;'"~`]|=(?:==?|>)?|\.+|\/[=*\/]?|\*[\/=]?|\+(?:=|\++)?|-(?:=|-+)?|[\^%]=?|&[&=]?|\|[\|=]?|>{1,3}=?|<<?=?|!={0,2}|(0|[1-9][0-9]*))(.*)$/;
-    var rx_digits = /^([0-9]+)(.*)$/;
-    var rx_hexs = /^([0-9a-fA-F]+)(.*)$/;
-    var rx_octals = /^([0-7]+)(.*)$/;
-    var rx_bits = /^([01]+)(.*)$/;
+        rx_token = /^((\s+)|([a-zA-Z_$][a-zA-Z0-9_$]*)|[(){}\[\]\?,:;'"~`]|=(?:==?|>)?|\.+|\/[=*\/]?|\*[\/=]?|\+(?:=|\++)?|-(?:=|-+)?|[\^%]=?|&[&=]?|\|[\|=]?|>{1,3}=?|<<?=?|!={0,2}|(0|[1-9][0-9]*))(.*)$/,
+        rx_digits = /^([0-9]+)(.*)$/,
+        rx_hexs = /^([0-9a-fA-F]+)(.*)$/,
+        rx_octals = /^([0-7]+)(.*)$/,
+        rx_bits = /^([01]+)(.*)$/,
 // mega
-    var rx_mega = /`|\$\{/;
+        rx_mega = /`|\$\{/,
 // indentation
-    var rx_colons = /^(.*)\?([:.]*)$/;
-    var rx_dot = /\.$/;
+        rx_colons = /^(.*)\?([:.]*)$/,
+        rx_dot = /\.$/,
 // JSON number
-    var rx_JSON_number = /^-?\d+(?:\.\d*)?(?:e[\-+]?\d+)?$/i;
-// initial cap
-    var rx_cap = /^[A-Z]/;
+        rx_JSON_number = /^-?\d+(?:\.\d*)?(?:e[\-+]?\d+)?$/i;
 
     function is_letter(string) {
-        return (string >= "a" && string <= "z\uffff") ||
-                (string >= "A" && string <= "Z\uffff");
+        return (string >= 'a' && string <= 'z\uffff') ||
+                (string >= 'A' && string <= 'Z\uffff');
     }
 
     function supplant(string, object) {
@@ -417,35 +414,34 @@ var jslint = (function JSLint() {
         });
     }
 
-    var anon = "anonymous"; // The guessed name for anonymous functions.
-    var blockage;           // The current block.
-    var block_stack;        // The stack of blocks.
-    var declared_globals;   // The object containing the global declarations.
-    var directives;         // The directive comments.
-    var directive_mode;     // true if directives are still allowed.
-    var early_stop;         // true if JSLint cannot finish.
-    var export_mode;        // true if an export statement was seen.
-    var fudge;              // true if the natural numbers start with 1.
-    var functionage;        // The current function.
-    var functions;          // The array containing all of the functions.
-    var global;             // The global object; the outermost context.
-    var imports;            // The array collecting all import-from strings.
-    var json_mode;          // true if parsing JSON.
-    var lines;              // The array containing source lines.
-    var module_mode;        // true if import or export was used.
-    var next_token;         // The next token to be examined in the parse.
-    var option;             // The options parameter.
-    var property;           // The object containing the tallied property names.
-    var mega_mode;          // true if currently parsing a megastring literal.
-    var stack;              // The stack of functions.
-    var syntax;             // The object containing the parser.
-    var token;              // The current token being examined in the parse.
-    var token_nr;           // The number of the next token.
-    var tokens;             // The array of tokens.
-    var tenure;             // The predefined property registry.
-    var tree;               // The abstract parse tree.
-    var var_mode;           // true if using var; false if using let.
-    var warnings;           // The array collecting all generated warnings.
+    var anon = "anonymous", // The guessed name for anonymous functions.
+        blockage,           // The current block.
+        block_stack,        // The stack of blocks.
+        declared_globals,   // The object containing the global declarations.
+        directive_mode,     // true if directives are still allowed.
+        early_stop,         // true if JSLint cannot finish.
+        export_mode,        // true if an export statement was seen.
+        fudge,              // true if the natural numbers start with 1.
+        functionage,        // The current function.
+        functions,          // The array containing all of the functions.
+        global,             // The global object, the outermost context.
+        imports,            // The array collecting all import-from strings.
+        json_mode,          // true if parsing JSON.
+        lines,              // The array containing source lines.
+        module_mode,        // true if import or export was used.
+        next_token,         // The next token to be examined in the parse.
+        option,             // The options parameter.
+        property,           // The object containing the tallied property names.
+        mega_mode,          // true if currently parsing a megastring literal.
+        stack,              // The stack of functions.
+        syntax,             // The object containing the parser.
+        token,              // The current token being examined in the parse.
+        token_nr,           // The number of the next token.
+        tokens,             // The array of tokens.
+        tenure,             // The predefined property registry.
+        tree,               // The abstract parse tree.
+        var_mode,           // true if using var, false if using let.
+        warnings;           // The array collecting all generated warnings.
 
 // Error reportage functions:
 
@@ -456,7 +452,7 @@ var jslint = (function JSLint() {
         if (the_token === undefined) {
             the_token = next_token;
         }
-        return (the_token.id === "(string)" || the_token.id === "(number)")
+        return (the_token.id === '(string)' || the_token.id === '(number)')
             ? String(the_token.value)
             : the_token.id;
     }
@@ -487,7 +483,7 @@ var jslint = (function JSLint() {
 // resembles an exception.
 
         var warning = {         // ~~
-            name: "JSLintError",
+            name: 'JSLintError',
             column: column,
             line: line,
             code: code
@@ -506,10 +502,9 @@ var jslint = (function JSLint() {
         }
         warning.message = supplant(bundle[code] || code, warning);
         warnings.push(warning);
-        return (
-            typeof option.maxerr === "number" &&
-            warnings.length === option.maxerr
-        )   ? stop_at("too_many", line, column)
+        return (typeof option.maxerr === 'number' &&
+                warnings.length === option.maxerr)
+            ? stop_at('too_many', line, column)
             : warning;
     }
 
@@ -552,7 +547,7 @@ var jslint = (function JSLint() {
         if (the_token === undefined) {
             the_token = next_token;
         }
-        delete the_token.warning;
+        the_token.warning = undefined;
         throw warn(code, the_token, a, b, c, d);
     }
 
@@ -575,16 +570,16 @@ var jslint = (function JSLint() {
             : source.split(rx_crlf);
         tokens = [];
 
-        var char;                   // a popular character
-        var column = 0;             // the column number of the next character
-        var from;                   // the starting column number of the token
-        var line = -1;              // the line number of the next character
-        var previous = global;      // the previous token including comments
-        var prior = global;         // the previous token excluding comments
-        var mega_from;              // the starting column of megastring
-        var mega_line;              // the starting line of megastring
-        var snippet;                // a piece of string
-        var source_line;            // the current line source string
+        var char,                   // a popular character
+            column = 0,             // the column number of the next character
+            from,                   // the starting column number of the token
+            line = -1,              // the line number of the next character
+            previous = global,      // the previous token including comments
+            prior = global,         // the previous token excluding comments
+            mega_from,              // the starting column of megastring
+            mega_line,              // the starting line of megastring
+            snippet,                // a piece of string
+            source_line;            // the current line source string
 
         function next_line() {
 
@@ -600,24 +595,24 @@ var jslint = (function JSLint() {
                 at = source_line.search(rx_tab);
                 if (at >= 0) {
                     if (!option.white) {
-                        warn_at("use_spaces", line, at + 1);
+                        warn_at('use_spaces', line, at + 1);
                     }
-                    source_line = source_line.replace(rx_tab, " ");
+                    source_line = source_line.replace(rx_tab, ' ');
                 }
                 at = source_line.search(rx_unsafe);
                 if (at >= 0) {
                     warn_at(
-                        "unsafe",
+                        'unsafe',
                         line,
                         column + at,
-                        "U+" + source_line.charCodeAt(at).toString(16)
+                        'U+' + source_line.charCodeAt(at).toString(16)
                     );
                 }
                 if (option.maxlen && option.maxlen < source_line.length) {
-                    warn_at("too_long", line, source_line.length);
-                } else if (!option.white && source_line.slice(-1) === " ") {
+                    warn_at('too_long', line, source_line.length);
+                } else if (!option.white && source_line.slice(-1) === ' ') {
                     warn_at(
-                        "unexpected_trailing_space",
+                        'unexpected_trailing_space',
                         line,
                         source_line.length - 1
                     );
@@ -647,15 +642,15 @@ var jslint = (function JSLint() {
 // matched an expected value.
 
             if (match !== undefined && char !== match) {
-                return stop_at("expected_a_b", line, column, match, char);
+                return stop_at('expected_a_b', line, column, match, char);
             }
             if (source_line) {
                 char = source_line.charAt(0);
                 source_line = source_line.slice(1);
                 snippet += char;
             } else {
-                char = "";
-                snippet += " ";
+                char = '';
+                snippet += ' ';
             }
             column += 1;
             return char;
@@ -672,7 +667,7 @@ var jslint = (function JSLint() {
                 column -= 1;
                 snip();
             } else {
-                char = "";
+                char = '';
             }
             return char;
         }
@@ -685,10 +680,10 @@ var jslint = (function JSLint() {
                 source_line = result[2];
                 snippet += char;
             } else {
-                char = "";
+                char = '';
                 if (!quiet) {
                     warn_at(
-                        "expected_digits_after_a",
+                        'expected_digits_after_a',
                         line,
                         column,
                         snippet
@@ -699,45 +694,44 @@ var jslint = (function JSLint() {
         }
 
         function escape(extra) {
-            switch (next_char("\\")) {
-            case "\\":
-            case "\"":
-            case "'":
-            case "/":
-            case ":":
-            case "=":
-            case "|":
-            case "b":
-            case "f":
-            case "n":
-            case "r":
-            case "t":
-            case " ":
+            switch (next_char('\\')) {
+            case '\\':
+            case '\'':
+            case '"':
+            case '/':
+            case ':':
+            case '=':
+            case '|':
+            case 'b':
+            case 'f':
+            case 'n':
+            case 'r':
+            case 't':
+            case ' ':
                 break;
-            case "u":
-                if (next_char("u") === "{") {
+            case 'u':
+                if (next_char('u') === '{') {
                     if (some_digits(rx_hexs) > 5) {
-                        warn_at("too_many_digits", line, column - 1);
+                        warn_at('too_many_digits', line, column - 1);
                     }
                     if (!option.es6) {
-                        warn_at("es6", line, column, "u{");
+                        warn_at('es6', line, column);
                     }
-                    if (next_char() !== "}") {
-                        stop_at("expected_a_before_b", line, column, "}", char);
+                    if (next_char() !== '}') {
+                        stop_at('expected_a_before_b', line, column, '}', char);
                     }
-                    next_char();
                     return;
                 }
                 back_char();
                 if (some_digits(rx_hexs, true) < 4) {
-                    warn_at("expected_four_digits", line, column - 1);
+                    warn_at('expected_four_digits', line, column - 1);
                 }
                 break;
-            case "":
-                return stop_at("unclosed_string", line, column);
+            case '':
+                return stop_at('unclosed_string', line, column);
             default:
                 if (extra && extra.indexOf(char) < 0) {
-                    warn_at("unexpected_a_after_b", line, column, char, "\\");
+                    warn_at('unexpected_a_after_b', line, column, char, '\\');
                 }
             }
             next_char();
@@ -758,7 +752,7 @@ var jslint = (function JSLint() {
 
 // Directives must appear before the first statement.
 
-            if (id !== "(comment)" && id !== ";") {
+            if (id !== '(comment)') {
                 directive_mode = false;
             }
 
@@ -769,7 +763,7 @@ var jslint = (function JSLint() {
             }
 
 // If this token is an identifier that touches a preceding number, or
-// a "/", comment, or regular expression literal that touches a preceding
+// a '/', comment, or regular expression literal that touches a preceding
 // comment or regular expression literal, then give a missing space warning.
 // This warning is not suppressed by option.white.
 
@@ -777,24 +771,24 @@ var jslint = (function JSLint() {
                 previous.line === line &&
                 previous.thru === from &&
                 (
-                    (id === "(comment)" || id === "(regexp)" || id === "/") &&
+                    (id === '(comment)' || id === '(regexp)' || id === '/') &&
                     (
-                        previous.id === "(comment)" ||
-                        previous.id === "(regexp)"
+                        previous.id === '(comment)' ||
+                        previous.id === '(regexp)'
                     )
                 )
             ) {
                 warn(
-                    "expected_space_a_b",
+                    'expected_space_a_b',
                     the_token,
                     artifact(previous),
                     artifact(the_token)
                 );
             }
-            if (previous.id === "." && id === "(number)") {
-                warn("expected_a_before_b", previous, "0", ".");
+            if (previous.id === '.' && id === '(number)') {
+                warn('expected_a_before_b', previous, '0', '.');
             }
-            if (prior.id === "." && the_token.identifier) {
+            if (prior.id === '.' && the_token.identifier) {
                 the_token.dot = true;
             }
 
@@ -803,16 +797,16 @@ var jslint = (function JSLint() {
             previous = the_token;
 
 // The prior token is a previous token that was not a comment. The prior token
-// is used to disambiguate "/", which can mean division or regular expression
+// is used to disambiguate '/', which can mean division or regular expression
 // literal.
 
-            if (previous.id !== "(comment)") {
+            if (previous.id !== '(comment)') {
                 prior = previous;
             }
             return the_token;
         }
 
-        function parse_directive(the_comment, body) {
+        function directive(the_comment, body) {
 
 // JSLint recognizes three directives that can be encoded in comments. This
 // function processes one item, and calls itself recursively to process the
@@ -820,68 +814,60 @@ var jslint = (function JSLint() {
 
             var result = body.match(rx_directive_part);
             if (result) {
-                var allowed;
-                var name = result[1];
-                var value = result[2];
+                var allowed,
+                    name = result[1],
+                    value = result[2];
                 switch (the_comment.directive) {
-                case "jslint":
+                case 'jslint':
                     allowed = allowed_option[name];
                     switch (typeof allowed) {
-                    case "boolean":
-                    case "object":
+                    case 'boolean':
+                    case 'object':
                         switch (value) {
-                        case "true":
-                        case "":
+                        case 'true':
+                        case '':
                         case undefined:
                             option[name] = true;
                             if (Array.isArray(allowed)) {
                                 populate(declared_globals, allowed, false);
                             }
                             break;
-                        case "false":
+                        case 'false':
                             option[name] = false;
                             break;
                         default:
-                            warn(
-                                "bad_option_a",
-                                the_comment,
-                                name + ":" + value
-                            );
+                            warn('bad_option_a', the_comment, name + ':' + value);
                         }
                         break;
-                    case "number":
+                    case 'number':
                         if (isFinite(+value)) {
                             option[name] = +value;
                         } else {
-                            warn(
-                                "bad_option_a",
-                                the_comment,
-                                name + ":" + value
-                            );
+                            warn('bad_option_a', the_comment, name + ':' + value);
                         }
                         break;
                     default:
-                        warn("bad_option_a", the_comment, name);
+                        warn('bad_option_a', the_comment, name);
                     }
                     break;
-                case "property":
+                case 'property':
                     if (tenure === undefined) {
                         tenure = empty();
                     }
                     tenure[name] = true;
                     break;
-                case "global":
+                case 'global':
                     if (value) {
-                        warn("bad_option_a", the_comment, name + ":" + value);
+                        warn('bad_option_a', the_comment, name + ':' + value);
                     }
                     declared_globals[name] = false;
                     module_mode = the_comment;
                     break;
                 }
-                return parse_directive(the_comment, result[3]);
+                return directive(the_comment, result[3]);
             }
             if (body) {
-                return stop("bad_directive_a", the_comment, body);
+                return stop('bad_directive_a', the_comment, body);
             }
         }
 
@@ -890,22 +876,21 @@ var jslint = (function JSLint() {
 // Make a comment object. Comments are not allowed in JSON text. Comments can
 // include directives and notices of incompletion.
 
-            var the_comment = make("(comment)", snippet);
+            var the_comment = make('(comment)', snippet);
             if (Array.isArray(snippet)) {
-                snippet = snippet.join(" ");
+                snippet = snippet.join(' ');
             }
             if (!option.devel && rx_todo.test(snippet)) {
-                warn("todo_comment", the_comment);
+                warn('todo_comment', the_comment);
             }
             var result = snippet.match(rx_directive);
             if (result) {
                 if (!directive_mode) {
-                    warn_at("misplaced_directive_a", line, from, result[1]);
+                    warn_at('misplaced_directive_a', line, from, result[1]);
                 } else {
                     the_comment.directive = result[1];
-                    parse_directive(the_comment, result[2]);
+                    directive(the_comment, result[2]);
                 }
-                directives.push(the_comment);
             }
             return the_comment;
         }
@@ -914,34 +899,35 @@ var jslint = (function JSLint() {
 
 // Parse a regular expression literal.
 
-            var result;
-            var value;
+            var result,
+                u_mode = false,
+                value;
 
             function quantifier() {
 
 // Match an optional quantifier.
 
                 switch (char) {
-                case "?":
-                case "*":
-                case "+":
+                case '?':
+                case '*':
+                case '+':
                     next_char();
                     break;
-                case "{":
+                case '{':
                     if (some_digits(rx_digits, true) === 0) {
-                        warn_at("expected_a", line, column, "0");
+                        warn_at('expected_a', line, column, '0');
                     }
-                    if (next_char() === ",") {
+                    if (next_char() === ',') {
                         some_digits(rx_digits, true);
                         next_char();
                     }
-                    next_char("}");
+                    next_char('}');
                     break;
                 default:
                     return;
                 }
-                if (char === "?") {
-                    next_char("?");
+                if (char === '?') {
+                    next_char('?');
                 }
             }
 
@@ -950,25 +936,25 @@ var jslint = (function JSLint() {
 // Match a character in a character class.
 
                 switch (char) {
-                case "\\":
+                case '\\':
                     escape();
                     return true;
-                case "[":
-                case "]":
-                case "/":
-                case "^":
-                case "-":
-                case "|":
-                case "":
+                case '[':
+                case ']':
+                case '/':
+                case '^':
+                case '-':
+                case '|':
+                case '':
                     return false;
-                case "`":
+                case '`':
                     if (mega_mode) {
-                        warn_at("unexpected_a", line, column, "`");
+                        warn_at('unexpected_a', line, column, '`');
                     }
                     next_char();
                     return true;
-                case " ":
-                    warn_at("expected_a_before_b", line, column, "\\", " ");
+                case ' ':
+                    warn_at('expected_a_before_b', line, column, '\\', ' ');
                     next_char();
                     return true;
                 default:
@@ -982,15 +968,10 @@ var jslint = (function JSLint() {
 // Match a range of subclasses.
 
                 if (subklass()) {
-                    if (char === "-") {
-                        next_char("-");
+                    if (char === '-') {
+                        next_char('-');
                         if (!subklass()) {
-                            return stop_at(
-                                "unexpected_a",
-                                line,
-                                column - 1,
-                                "-"
-                            );
+                            return stop_at('unexpected_a', line, column - 1, '-');
                         }
                     }
                     return range();
@@ -1001,12 +982,12 @@ var jslint = (function JSLint() {
 
 // Match a class.
 
-                next_char("[");
-                if (char === "^") {
-                    next_char("^");
+                next_char('[');
+                if (char === '^') {
+                    next_char('^');
                 }
                 range();
-                next_char("]");
+                next_char(']');
             }
 
             function choice() {
@@ -1015,54 +996,54 @@ var jslint = (function JSLint() {
 
 // Match a group that starts with left paren.
 
-                    next_char("(");
-                    if (char === "?") {
-                        next_char("?");
+                    next_char('(');
+                    if (char === '?') {
+                        next_char('?');
                         switch (char) {
-                        case ":":
-                        case "=":
-                        case "!":
+                        case ':':
+                        case '=':
+                        case '!':
                             next_char();
                             break;
                         default:
-                            next_char(":");
+                            next_char(':');
                         }
-                    } else if (char === ":") {
-                        warn_at("expected_a_before_b", line, column, "?", ":");
+                    } else if (char === ':') {
+                        warn_at('expected_a_before_b', line, column, '?', ':');
                     }
                     choice();
-                    next_char(")");
+                    next_char(')');
                 }
 
                 function factor() {
                     switch (char) {
-                    case "[":
+                    case '[':
                         klass();
                         return true;
-                    case "\\":
-                        escape("BbDdSsWw^${}[]().|*+?");
+                    case '\\':
+                        escape('BbDdSsWw^${}[]().|*+?');
                         return true;
-                    case "(":
+                    case '(':
                         group();
                         return true;
-                    case "/":
-                    case "|":
-                    case "]":
-                    case ")":
-                    case "}":
-                    case "{":
-                    case "?":
-                    case "+":
-                    case "*":
-                    case "":
+                    case '/':
+                    case '|':
+                    case ']':
+                    case ')':
+                    case '}':
+                    case '{':
+                    case '?':
+                    case '+':
+                    case '*':
+                    case '':
                         return false;
-                    case "`":
+                    case '`':
                         if (mega_mode) {
-                            warn_at("unexpected_a", line, column, "`");
+                            warn_at('unexpected_a', line, column, '`');
                         }
                         break;
-                    case " ":
-                        warn_at("expected_a_before_b", line, column, "\\", " ");
+                    case ' ':
+                        warn_at('expected_a_before_b', line, column, '\\', ' ');
                         break;
                     }
                     next_char();
@@ -1075,15 +1056,15 @@ var jslint = (function JSLint() {
                         return sequence(true);
                     }
                     if (!follow) {
-                        warn_at("expected_regexp_factor_a", line, column, char);
+                        warn_at('expected_regexp_factor_a', line, column, char);
                     }
                 }
 
 // Match a choice (a sequence that can be followed by | and another choice).
 
                 sequence();
-                if (char === "|") {
-                    next_char("|");
+                if (char === '|') {
+                    next_char('|');
                     return choice();
                 }
             }
@@ -1091,10 +1072,10 @@ var jslint = (function JSLint() {
 // Scan the regexp literal. Give a warning if the first character is = because
 // /= looks like a division assignment operator.
 
-            snippet = "";
+            snippet = '';
             next_char();
-            if (char === "=") {
-                warn_at("expected_a_before_b", line, column, "\\", "=");
+            if (char === '=') {
+                warn_at('expected_a_before_b', line, column, '\\', '=');
             }
             choice();
 
@@ -1102,30 +1083,22 @@ var jslint = (function JSLint() {
 
             snip();
             value = snippet;
-            next_char("/");
+            next_char('/');
 
 // Process dangling flag letters.
 
             var allowed = {
-                g: true,
-                i: true,
-                m: true,
-                u: 6,
-                y: 6
-            };
-            var flag = empty();
+                    g: true,
+                    i: true,
+                    m: true,
+                    u: u_mode,
+                    y: option.es6
+                },
+                flag = empty();
             (function make_flag() {
                 if (is_letter(char)) {
-                    switch (allowed[char]) {
-                    case true:
-                        break;
-                    case 6:
-                        if (!option.es6) {
-                            warn_at("es6", line, column, char);
-                        }
-                        break;
-                    default:
-                        warn_at("unexpected_a", line, column, char);
+                    if (allowed[char] !== true) {
+                        warn_at('unexpected_a', line, column, char);
                     }
                     allowed[char] = false;
                     flag[char] = true;
@@ -1133,11 +1106,14 @@ var jslint = (function JSLint() {
                     return make_flag();
                 }
             }());
-            back_char();
-            if (char === "/" || char === "*") {
-                return stop_at("unexpected_a", line, from, char);
+            if (u_mode && !flag.u) {
+                warn_at('expected_a_before_b', line, column, 'u', char);
             }
-            result = make("(regexp)", char);
+            back_char();
+            if (char === '/' || char === '*') {
+                return stop_at('unexpected_a', line, from, char);
+            }
+            result = make('(regexp)', char);
             result.flag = flag;
             result.value = value;
             return result;
@@ -1148,26 +1124,26 @@ var jslint = (function JSLint() {
 // Make a string token.
 
             var the_token;
-            snippet = "";
+            snippet = '';
             next_char();
 
             return (function next() {
                 switch (char) {
                 case quote:
                     snip();
-                    the_token = make("(string)", snippet);
+                    the_token = make('(string)', snippet);
                     the_token.quote = quote;
                     return the_token;
-                case "\\":
+                case '\\':
                     escape();
                     break;
-                case "":
-                    return stop_at("unclosed_string", line, column);
-                case "`":
+                case '':
+                    return stop_at('unclosed_string', line, column);
+                case '`':
                     if (mega_mode) {
-                        warn_at("unexpected_a", line, column, "`");
+                        warn_at('unexpected_a', line, column, '`');
                     }
-                    next_char("`");
+                    next_char('`');
                     break;
                 default:
                     next_char();
@@ -1177,13 +1153,13 @@ var jslint = (function JSLint() {
         }
 
         function frack() {
-            if (char === ".") {
+            if (char === '.') {
                 some_digits(rx_digits);
                 next_char();
             }
-            if (char === "E" || char === "e") {
+            if (char === 'E' || char === 'e') {
                 next_char();
-                if (char !== "+" && char !== "-") {
+                if (char !== '+' && char !== '-') {
                     back_char();
                 }
                 some_digits(rx_digits);
@@ -1192,20 +1168,20 @@ var jslint = (function JSLint() {
         }
 
         function number() {
-            if (snippet === "0") {
+            if (snippet === '0') {
                 switch (next_char()) {
-                case ".":
+                case '.':
                     frack();
                     break;
-                case "b":
+                case 'b':
                     some_digits(rx_bits);
                     next_char();
                     break;
-                case "o":
+                case 'o':
                     some_digits(rx_octals);
                     next_char();
                     break;
-                case "x":
+                case 'x':
                     some_digits(rx_hexs);
                     next_char();
                     break;
@@ -1219,12 +1195,12 @@ var jslint = (function JSLint() {
 // unexpected is going on.
 
             if (
-                (char >= "0" && char <= "9") ||
-                (char >= "a" && char <= "z") ||
-                (char >= "A" && char <= "Z")
+                (char >= '0' && char <= '9') ||
+                (char >= 'a' && char <= 'z') ||
+                (char >= 'A' && char <= 'Z')
             ) {
                 return stop_at(
-                    "unexpected_a_after_b",
+                    'unexpected_a_after_b',
                     line,
                     column - 1,
                     snippet.slice(-1),
@@ -1232,23 +1208,23 @@ var jslint = (function JSLint() {
                 );
             }
             back_char();
-            return make("(number)", snippet);
+            return make('(number)', snippet);
         }
 
         function lex() {
-            var array;
-            var i = 0;
-            var j = 0;
-            var last;
-            var result;
-            var the_token;
+            var array,
+                i = 0,
+                j = 0,
+                last,
+                result,
+                the_token;
             if (!source_line) {
                 source_line = next_line();
                 from = 0;
                 return (source_line === undefined)
                     ? (mega_mode)
-                        ? stop_at("unclosed_mega", mega_line, mega_from)
-                        : make("(end)")
+                        ? stop_at('unclosed_mega', mega_line, mega_from)
+                        : make('(end)')
                     : lex();
             }
             from = column;
@@ -1261,12 +1237,7 @@ var jslint = (function JSLint() {
 // result[5] rest
 
             if (!result) {
-                return stop_at(
-                    "unexpected_char_a",
-                    line,
-                    column,
-                    source_line.charAt(0)
-                );
+                return stop_at('unexpected_char_a', line, column, source_line.charAt(0));
             }
 
             snippet = result[1];
@@ -1297,24 +1268,24 @@ var jslint = (function JSLint() {
 
 // The token is a single or double quote string.
 
-            case "\"":
-            case "'":
+            case '\'':
+            case '"':
                 return string(snippet);
 
-// The token is a megastring. We don't allow any kind of mega nesting.
+// The token is a megastring. We don't allow any kind if mega nesting.
 
-            case "`":
+            case '`':
                 if (mega_mode) {
-                    return stop_at("expected_a_b", line, column, "}", "`");
+                    return stop_at('expected_a_b', line, column, '}', '`');
                 }
-                snippet = "";
+                snippet = '';
                 mega_from = from;
                 mega_line = line;
                 mega_mode = true;
 
 // Parsing a mega literal is tricky. First make a ` token.
 
-                make("`");
+                make('`');
                 from += 1;
 
 // Then loop, building up a string, possibly from many lines, until seeing
@@ -1327,9 +1298,9 @@ var jslint = (function JSLint() {
 // If neither ` nor ${ is seen, then the whole line joins the snippet.
 
                     if (at < 0) {
-                        snippet += source_line + "\n";
+                        snippet += source_line + '\n';
                         return (next_line() === undefined)
-                            ? stop_at("unclosed_mega", mega_line, mega_from)
+                            ? stop_at('unclosed_mega', mega_line, mega_from)
                             : part();
                     }
 
@@ -1339,28 +1310,28 @@ var jslint = (function JSLint() {
                     snippet += source_line.slice(0, at);
                     column += at;
                     source_line = source_line.slice(at);
-                    make("(string)", snippet).quote = "`";
-                    snippet = "";
+                    make('(string)', snippet).quote = '`';
+                    snippet = '';
 
 // If ${, then make tokens that will become part of an expression until
 // a } token is made.
 
-                    if (source_line.charAt(0) === "$") {
+                    if (source_line.charAt(0) === '$') {
                         column += 2;
-                        make("${");
+                        make('${');
                         source_line = source_line.slice(2);
                         (function expr() {
                             var id = lex().id;
-                            if (id === "{") {
+                            if (id === '{') {
                                 return stop_at(
-                                    "expected_a_b",
+                                    'expected_a_b',
                                     line,
                                     column,
-                                    "}",
-                                    "{"
+                                    '}',
+                                    '{'
                                 );
                             }
-                            if (id !== "}") {
+                            if (id !== '}') {
                                 return expr();
                             }
                         }());
@@ -1370,48 +1341,48 @@ var jslint = (function JSLint() {
                 source_line = source_line.slice(1);
                 column += 1;
                 mega_mode = false;
-                return make("`");
+                return make('`');
 
 // The token is a // comment.
 
-            case "//":
+            case '//':
                 snippet = source_line;
-                source_line = "";
+                source_line = '';
                 the_token = comment(snippet);
                 if (mega_mode) {
-                    warn("unexpected_comment", the_token, "`");
+                    warn('unexpected_comment', the_token, '`');
                 }
                 return the_token;
 
 // The token is a /* comment.
 
-            case "/*":
+            case '/*':
                 array = [];
-                if (source_line.charAt(0) === "/") {
-                    warn_at("unexpected_a", line, column + i, "/");
+                if (source_line.charAt(0) === '/') {
+                    warn_at('unexpected_a', line, column + i, '/');
                 }
                 (function next() {
-                    if (source_line > "") {
+                    if (source_line > '') {
                         i = source_line.search(rx_star_slash);
                         if (i >= 0) {
                             return;
                         }
                         j = source_line.search(rx_slash_star);
                         if (j >= 0) {
-                            warn_at("nested_comment", line, column + j);
+                            warn_at('nested_comment', line, column + j);
                         }
                     }
                     array.push(source_line);
                     source_line = next_line();
                     if (source_line === undefined) {
-                        return stop_at("unclosed_comment", line, column);
+                        return stop_at('unclosed_comment', line, column);
                     }
                     return next();
                 }());
                 snippet = source_line.slice(0, i);
                 j = snippet.search(rx_slash_star_or_slash);
                 if (j >= 0) {
-                    warn_at("nested_comment", line, column + j);
+                    warn_at('nested_comment', line, column + j);
                 }
                 array.push(snippet);
                 column += i + 2;
@@ -1420,13 +1391,13 @@ var jslint = (function JSLint() {
 
 // The token is a slash.
 
-            case "/":
+            case '/':
 
 // The / can be a division operator or the beginning of a regular expression
 // literal. It is not possible to know which without doing a complete parse.
 // We want to complete the tokenization before we begin to parse, so we will
 // estimate. This estimator can fail in some cases. For example, it cannot
-// know if "}" is ending a block or ending an object literal, so it can
+// know if '}' is ending a block or ending an object literal, so it can
 // behave incorrectly in that case; it is not meaningful to divide an
 // object, so it is likely that we can get away with it. We avoided the worst
 // cases by eliminating automatic semicolon insertion.
@@ -1434,37 +1405,37 @@ var jslint = (function JSLint() {
                 if (prior.identifier) {
                     if (!prior.dot) {
                         switch (prior.id) {
-                        case "return":
+                        case 'return':
                             return regexp();
-                        case "(begin)":
-                        case "case":
-                        case "delete":
-                        case "in":
-                        case "instanceof":
-                        case "new":
-                        case "typeof":
-                        case "void":
-                        case "yield":
+                        case '(begin)':
+                        case 'case':
+                        case 'delete':
+                        case 'in':
+                        case 'instanceof':
+                        case 'new':
+                        case 'typeof':
+                        case 'void':
+                        case 'yield':
                             the_token = regexp();
-                            return stop("unexpected_a", the_token);
+                            return stop('unexpected_a', the_token);
                         }
                     }
                 } else {
                     last = prior.id.charAt(prior.id.length - 1);
-                    if ("(,=:?[".indexOf(last) >= 0) {
+                    if ('(,=:?['.indexOf(last) >= 0) {
                         return regexp();
                     }
-                    if ("!&|{};~+-*%/^<>".indexOf(last) >= 0) {
+                    if ('!&|{};~+-*%/^<>'.indexOf(last) >= 0) {
                         the_token = regexp();
-                        warn("wrap_regexp", the_token);
+                        warn('wrap_regexp', the_token);
                         return the_token;
                     }
                 }
-                if (source_line.charAt(0) === "/") {
+                if (source_line.charAt(0) === '/') {
                     column += 1;
                     source_line = source_line.slice(1);
-                    snippet = "/=";
-                    warn_at("unexpected_a", line, column, "/=");
+                    snippet = '/=';
+                    warn_at('unexpected_a', line, column, '/=');
                 }
                 break;
             }
@@ -1475,7 +1446,7 @@ var jslint = (function JSLint() {
 // when ES6 has been finished and widely deployed and adopted.
 
         while (true) {
-            if (lex().id === "(end)") {
+            if (lex().id === '(end)') {
                 break;
             }
         }
@@ -1501,20 +1472,20 @@ var jslint = (function JSLint() {
 // Tally the property name. If it is a string, only tally strings that conform
 // to the identifier rules.
 
-        if (id === "(string)") {
+        if (id === '(string)') {
             id = name.value;
             if (!rx_identifier.test(id)) {
                 return id;
             }
         } else {
             if (!name.identifier) {
-                return stop("expected_identifier_a", name);
+                return stop('expected_identifier_a', name);
             }
         }
 
 // If we have seen this name before, increment its count.
 
-        if (typeof property[id] === "number") {
+        if (typeof property[id] === 'number') {
             property[id] += 1;
 
 // If this is the first time seeing this property name, and if there is a
@@ -1524,11 +1495,11 @@ var jslint = (function JSLint() {
         } else {
             if (tenure !== undefined) {
                 if (tenure[id] !== true) {
-                    warn("unregistered_property_a", name);
+                    warn('unregistered_property_a', name);
                 }
             } else {
                 if (name.identifier && rx_bad_property.test(id)) {
-                    warn("bad_property_a", name);
+                    warn('bad_property_a', name);
                 }
             }
             property[id] = 1;
@@ -1542,9 +1513,9 @@ var jslint = (function JSLint() {
 
         var cadet = tokens[token_nr];
         token_nr += 1;
-        if (cadet.id === "(comment)") {
+        if (cadet.id === '(comment)') {
             if (json_mode) {
-                warn("unexpected_a", cadet);
+                warn('unexpected_a', cadet);
             }
             return dispense();
         } else {
@@ -1556,8 +1527,8 @@ var jslint = (function JSLint() {
 
 // Look ahead one token without advancing.
 
-        var old_token_nr = token_nr;
-        var cadet = dispense(true);
+        var old_token_nr = token_nr,
+            cadet = dispense(true);
         token_nr = old_token_nr;
         return cadet;
     }
@@ -1568,9 +1539,9 @@ var jslint = (function JSLint() {
 
 // Attempt to give helpful names to anonymous functions.
 
-        if (token.identifier && token.id !== "function") {
+        if (token.identifier && token.id !== 'function') {
             anon = token.id;
-        } else if (token.id === "(string)" && rx_identifier.test(token.value)) {
+        } else if (token.id === '(string)' && rx_identifier.test(token.value)) {
             anon = token.value;
         }
 
@@ -1578,9 +1549,9 @@ var jslint = (function JSLint() {
 
         if (id !== undefined && next_token.id !== id) {
             return (match === undefined)
-                ? stop("expected_a_b", next_token, id, artifact())
+                ? stop('expected_a_b', next_token, id, artifact())
                 : stop(
-                    "expected_a_b_from_c_d",
+                    'expected_a_b_from_c_d',
                     next_token,
                     id,
                     artifact(match),
@@ -1593,7 +1564,7 @@ var jslint = (function JSLint() {
 
         token = next_token;
         next_token = dispense();
-        if (next_token.id === "(end)") {
+        if (next_token.id === '(end)') {
             token_nr -= 1;
         }
     }
@@ -1603,78 +1574,78 @@ var jslint = (function JSLint() {
     function json_value() {
 
         function json_object() {
-            var brace = next_token;
-            var object = empty();
-            advance("{");
-            if (next_token.id !== "}") {
+            var brace = next_token,
+                object = empty();
+            advance('{');
+            if (next_token.id !== '}') {
                 (function next() {
-                    if (next_token.quote !== "\"") {
-                        warn("unexpected_a", next_token, next_token.quote);
+                    if (next_token.quote !== '"') {
+                        warn('unexpected_a', next_token, next_token.quote);
                     }
-                    advance("(string)");
+                    advance('(string)');
                     if (object[token.value] !== undefined) {
-                        warn("duplicate_a", token);
-                    } else if (token.value === "__proto__") {
-                        warn("bad_property_name_a", token);
+                        warn('duplicate_a', token);
+                    } else if (token.value === '__proto__') {
+                        warn('bad_property_name_a', token);
                     } else {
                         object[token.value] = token;
                     }
-                    advance(":");
+                    advance(':');
                     json_value();
-                    if (next_token.id === ",") {
-                        advance(",");
+                    if (next_token.id === ',') {
+                        advance(',');
                         return next();
                     }
                 }());
             }
-            advance("}", brace);
+            advance('}', brace);
         }
 
         function json_array() {
             var bracket = next_token;
-            advance("[");
-            if (next_token.id !== "]") {
+            advance('[');
+            if (next_token.id !== ']') {
                 (function next() {
                     json_value();
-                    if (next_token.id === ",") {
-                        advance(",");
+                    if (next_token.id === ',') {
+                        advance(',');
                         return next();
                     }
                 }());
             }
-            advance("]", bracket);
+            advance(']', bracket);
         }
 
         switch (next_token.id) {
-        case "{":
+        case '{':
             json_object();
             break;
-        case "[":
+        case '[':
             json_array();
             break;
-        case "true":
-        case "false":
-        case "null":
+        case 'true':
+        case 'false':
+        case 'null':
             advance();
             break;
-        case "(number)":
+        case '(number)':
             if (!rx_JSON_number.test(next_token.value)) {
-                warn("unexpected_a");
+                warn('unexpected_a');
             }
             advance();
             break;
-        case "(string)":
-            if (next_token.quote !== "\"") {
-                warn("unexpected_a", next_token, next_token.quote);
+        case '(string)':
+            if (next_token.quote !== '"') {
+                warn('unexpected_a', next_token, next_token.quote);
             }
             advance();
             break;
-        case "-":
-            advance("-");
-            advance("(number)");
+        case '-':
+            advance('-');
+            advance('(number)');
             break;
         default:
-            stop("unexpected_a");
+            stop('unexpected_a');
         }
     }
 
@@ -1690,8 +1661,8 @@ var jslint = (function JSLint() {
 
 // Reserved words may not be enrolled.
 
-        if (syntax[id] !== undefined && id !== "ignore") {
-            warn("reserved_a", name);
+        if (syntax[id] !== undefined && id !== 'ignore') {
+            warn('reserved_a', name);
         } else {
 
 // Has the name been enrolled in this context?
@@ -1699,7 +1670,7 @@ var jslint = (function JSLint() {
             var earlier = functionage.context[id];
             if (earlier) {
                 warn(
-                    "redefinition_a_b",
+                    'redefinition_a_b',
                     name,
                     name.id,
                     earlier.line + fudge
@@ -1715,17 +1686,17 @@ var jslint = (function JSLint() {
                     }
                 });
                 if (earlier) {
-                    if (id === "ignore") {
-                        if (earlier.role === "variable") {
-                            warn("unexpected_a", name);
+                    if (id === 'ignore') {
+                        if (earlier.role === 'variable') {
+                            warn('unexpected_a', name);
                         }
                     } else {
                         if ((
-                            role !== "exception" ||
-                            earlier.role !== "exception"
-                        ) && role !== "parameter") {
+                            role !== 'exception' ||
+                            earlier.role !== 'exception'
+                        ) && role !== 'parameter') {
                             warn(
-                                "redefinition_a_b",
+                                'redefinition_a_b',
                                 name,
                                 name.id,
                                 earlier.line + fudge
@@ -1761,8 +1732,7 @@ var jslint = (function JSLint() {
 // process leds (infix operators) until the bind powers cause it to stop. It
 // returns the expression's parse tree.
 
-        var left;
-        var the_symbol;
+        var left, the_symbol;
 
 // Statements will have already advanced, so advance now only if the token is
 // not the first of a statement,
@@ -1775,9 +1745,9 @@ var jslint = (function JSLint() {
             left = the_symbol.nud();
         } else if (token.identifier) {
             left = token;
-            left.arity = "variable";
+            left.arity = 'variable';
         } else {
-            return stop("unexpected_a", token);
+            return stop('unexpected_a', token);
         }
         (function right() {
             the_symbol = syntax[next_token.id];
@@ -1798,33 +1768,33 @@ var jslint = (function JSLint() {
 
 // Parse the condition part of a do, if, while.
 
-        var the_paren = next_token;
-        var the_value;
+        var the_paren = next_token,
+            the_value;
         the_paren.free = true;
-        advance("(");
+        advance('(');
         the_value = expression(0);
-        advance(")");
+        advance(')');
         if (the_value.wrapped === true) {
-            warn("unexpected_a", the_paren);
+            warn('unexpected_a', the_paren);
         }
         switch (the_value.id) {
-        case "?":
-        case "~":
-        case "&":
-        case "|":
-        case "^":
-        case "<<":
-        case ">>":
-        case ">>>":
-        case "+":
-        case "-":
-        case "*":
-        case "/":
-        case "%":
-        case "typeof":
-        case "(number)":
-        case "(string)":
-            warn("unexpected_a", the_value);
+        case '?':
+        case '~':
+        case '&':
+        case '|':
+        case '^':
+        case '<<':
+        case '>>':
+        case '>>>':
+        case '+':
+        case '-':
+        case '*':
+        case '/':
+        case '%':
+        case 'typeof':
+        case '(number)':
+        case '(string)':
+            warn('unexpected_a', the_value);
             break;
         }
         return the_value;
@@ -1832,11 +1802,11 @@ var jslint = (function JSLint() {
 
     function is_weird(thing) {
         return (
-            thing.id === "(regexp)" ||
-            thing.id === "{" ||
-            thing.id === "=>" ||
-            thing.id === "function" ||
-            (thing.id === "[" && thing.arity === "unary")
+            thing.id === '(regexp)' ||
+            thing.id === '{' ||
+            thing.id === '=>' ||
+            thing.id === 'function' ||
+            (thing.id === '[' && thing.arity === 'unary')
         );
     }
 
@@ -1856,49 +1826,44 @@ var jslint = (function JSLint() {
         if (Array.isArray(b)) {
             return false;
         }
-        if (a.id === "(number)" && b.id === "(number)") {
+        if (a.id === '(number)' && b.id === '(number)') {
             return a.value === b.value;
         }
-        var a_string;
-        var b_string;
-        if (a.id === "(string)") {
+        var a_string, b_string;
+        if (a.id === '(string)') {
             a_string = a.value;
-        } else if (a.id === "`" && a.constant) {
+        } else if (a.id === '`' && a.constant) {
             a_string = a.value[0];
         }
-        if (b.id === "(string)") {
+        if (b.id === '(string)') {
             b_string = b.value;
-        } else if (b.id === "`" && b.constant) {
+        } else if (b.id === '`' && b.constant) {
             b_string = b.value[0];
         }
-        if (typeof a_string === "string") {
+        if (typeof a_string === 'string') {
             return a_string === b_string;
         }
         if (is_weird(a) || is_weird(b)) {
             return false;
         }
         if (a.arity === b.arity && a.id === b.id) {
-            if (a.id === ".") {
+            if (a.id === '.') {
                 return are_similar(a.expression, b.expression) &&
                         are_similar(a.name, b.name);
             }
             switch (a.arity) {
-            case "unary":
+            case 'unary':
                 return are_similar(a.expression, b.expression);
-            case "binary":
-                return (
-                    a.id !== "(" &&
-                    are_similar(a.expression[0], b.expression[0]) &&
-                    are_similar(a.expression[1], b.expression[1])
-                );
-            case "ternary":
-                return (
-                    are_similar(a.expression[0], b.expression[0]) &&
-                    are_similar(a.expression[1], b.expression[1]) &&
-                    are_similar(a.expression[2], b.expression[2])
-                );
-            case "function":
-            case "regexp":
+            case 'binary':
+                return a.id !== '(' &&
+                        are_similar(a.expression[0], b.expression[0]) &&
+                        are_similar(a.expression[1], b.expression[1]);
+            case 'ternary':
+                return are_similar(a.expression[0], b.expression[0]) &&
+                        are_similar(a.expression[1], b.expression[1]) &&
+                        are_similar(a.expression[2], b.expression[2]);
+            case 'function':
+            case 'regexp':
                 return false;
             default:
                 return true;
@@ -1911,18 +1876,18 @@ var jslint = (function JSLint() {
 
 // Try to match a semicolon.
 
-        if (next_token.id === ";") {
-            advance(";");
+        if (next_token.id === ';') {
+            advance(';');
         } else {
             warn_at(
-                "expected_a_b",
+                'expected_a_b',
                 token.line,
                 token.thru,
-                ";",
+                ';',
                 artifact(next_token)
             );
         }
-        anon = "anonymous";
+        anon = 'anonymous';
     }
 
     function statement() {
@@ -1931,23 +1896,23 @@ var jslint = (function JSLint() {
 // have use for one. A statement can be one of the standard statements, or
 // an assignment expression, or an invocation expression.
 
-        var first;
-        var the_label;
-        var the_statement;
-        var the_symbol;
+        var first,
+            the_label,
+            the_statement,
+            the_symbol;
         advance();
-        if (token.identifier && next_token.id === ":") {
+        if (token.identifier && next_token.id === ':') {
             the_label = token;
-            if (the_label.id === "ignore") {
-                warn("unexpected_a", the_label);
+            if (the_label.id === 'ignore') {
+                warn('unexpected_a', the_label);
             }
-            advance(":");
+            advance(':');
             switch (next_token.id) {
-            case "do":
-            case "for":
-            case "switch":
-            case "while":
-                enroll(the_label, "label", true);
+            case 'do':
+            case 'for':
+            case 'switch':
+            case 'while':
+                enroll(the_label, 'label', true);
                 the_label.init = true;
                 the_label.dead = false;
                 the_statement = statement();
@@ -1956,7 +1921,7 @@ var jslint = (function JSLint() {
                 return the_statement;
             default:
                 advance();
-                warn("unexpected_label_a", the_label);
+                warn('unexpected_label_a', the_label);
             }
         }
 
@@ -1974,8 +1939,8 @@ var jslint = (function JSLint() {
 // It is an expression statement.
 
             the_statement = expression(0, true);
-            if (the_statement.wrapped && the_statement.id !== "(") {
-                warn("unexpected_a", first);
+            if (the_statement.wrapped && the_statement.id !== '(') {
+                warn('unexpected_a', first);
             }
             semicolon();
         }
@@ -1994,17 +1959,17 @@ var jslint = (function JSLint() {
         (function next(disrupt) {
             var a_statement;
             switch (next_token.id) {
-            case "}":
-            case "case":
-            case "default":
-            case "else":
-            case "(end)":
+            case '}':
+            case 'case':
+            case 'default':
+            case 'else':
+            case '(end)':
                 break;
             default:
                 a_statement = statement();
                 array.push(a_statement);
                 if (disrupt) {
-                    warn("unreachable_a", a_statement);
+                    warn('unreachable_a', a_statement);
                 }
                 return next(a_statement.disrupt);
             }
@@ -2017,7 +1982,7 @@ var jslint = (function JSLint() {
 // Some features should not be at the outermost level.
 
         if (functionage === global) {
-            warn("unexpected_at_top_level_a", thing);
+            warn('unexpected_at_top_level_a', thing);
         }
     }
 
@@ -2026,44 +1991,45 @@ var jslint = (function JSLint() {
 // Some features must be at the most outermost level.
 
         if (blockage !== global) {
-            warn("misplaced_a", the_thing);
+            warn('misplaced_a', the_thing);
         }
     }
 
     function block(special) {
 
 // Parse a block, a sequence of statements wrapped in braces.
-//  special "body"      The block is a function body.
-//          "ignore"    No warning on an empty block.
-//          "naked"     No advance.
+//  special 'body'      The block is a function body.
+//          'ignore'    No warning on an empty block.
+//          'naked'     No advance.
 //          undefined   Not special.
 
-        var stmts;
-        var the_block;
-        if (special !== "naked") {
-            advance("{");
+        var stmts, the_block;
+        if (special !== 'naked') {
+            advance('{');
         }
         the_block = token;
-        the_block.arity = "statement";
-        the_block.body = special === "body";
+        the_block.arity = 'statement';
+        the_block.body = special === 'body';
 
-// All top level function bodies should include the "use strict" pragma unless
+// All top level function bodies should include the 'use strict' pragma unless
 // the whole file is strict.
 
         if (the_block.body && stack.length <= 1 && !global.strict) {
             if (
-                next_token.id === "(string)" ||
-                next_token.value === "use strict"
+                next_token.id === '(string)' ||
+                next_token.value === 'use strict'
             ) {
                 next_token.statement = true;
                 functionage.strict = true;
-                advance("(string)");
-                advance(";");
+                advance('(string)');
+                advance(';');
             } else {
                 warn(
-                    "expected_a_before_b",
+                    'expected_a_before_b',
                     next_token,
-                    " \"use strict\"; ",
+                    (next_token.id === '`')
+                        ? '\''
+                        : 'use strict',
                     artifact(next_token)
                 );
             }
@@ -2071,14 +2037,14 @@ var jslint = (function JSLint() {
         stmts = statements();
         the_block.block = stmts;
         if (stmts.length === 0) {
-            if (!option.devel && special !== "ignore") {
-                warn("empty_block", the_block);
+            if (!option.devel && special !== 'ignore') {
+                warn('empty_block', the_block);
             }
             the_block.disrupt = false;
         } else {
             the_block.disrupt = stmts[stmts.length - 1].disrupt;
         }
-        advance("}");
+        advance('}');
         return the_block;
     }
 
@@ -2090,11 +2056,11 @@ var jslint = (function JSLint() {
 //      v
 
         if (
-            the_thing.id !== "." &&
-            (the_thing.id !== "[" || the_thing.arity !== "binary") &&
-            the_thing.arity !== "variable"
+            the_thing.id !== '.' &&
+            (the_thing.id !== '[' || the_thing.arity !== 'binary') &&
+            the_thing.arity !== 'variable'
         ) {
-            warn("bad_assignment_a", the_thing);
+            warn('bad_assignment_a', the_thing);
             return false;
         }
         return true;
@@ -2112,11 +2078,11 @@ var jslint = (function JSLint() {
         if (
             !left.identifier &&
             (
-                left.arity !== "binary" ||
-                (id !== "." && id !== "(" && id !== "[")
+                left.arity !== 'binary' ||
+                (id !== '.' && id !== '(' && id !== '[')
             )
         ) {
-            warn("unexpected_a", right);
+            warn('unexpected_a', right);
             return false;
         }
         return true;
@@ -2147,29 +2113,29 @@ var jslint = (function JSLint() {
 
         var the_symbol = symbol(id, 20);
         the_symbol.led = function (left) {
-            var the_token = token;
-            var right;
-            the_token.arity = "assignment";
+            var the_token = token,
+                right;
+            the_token.arity = 'assignment';
             right = expression(20 - 1);
-            if (id === "=" && left.arity === "variable") {
+            if (id === '=' && left.arity === 'variable') {
                 the_token.names = left;
                 the_token.expression = right;
             } else {
                 the_token.expression = [left, right];
             }
             switch (right.arity) {
-            case "assignment":
-            case "pre":
-            case "post":
-                warn("unexpected_a", right);
+            case 'assignment':
+            case 'pre':
+            case 'post':
+                warn('unexpected_a', right);
                 break;
             }
             if (
                 option.es6 &&
-                left.arity === "unary" &&
-                (left.id === "[" || left.id === "{")
+                left.arity === 'unary' &&
+                (left.id === '[' || left.id === '{')
             ) {
-                warn("expected_a_before_b", left, "const", left.id);
+                warn('expected_a_before_b', left, 'const', left.id);
             } else {
                 mutation_check(left);
             }
@@ -2183,8 +2149,7 @@ var jslint = (function JSLint() {
 // Make a constant symbol.
 
         var the_symbol = symbol(id);
-        the_symbol.constant = true;
-        the_symbol.nud = (typeof value === "function")
+        the_symbol.nud = (typeof value === 'function')
             ? value
             : function () {
                 token.constant = true;
@@ -2205,7 +2170,7 @@ var jslint = (function JSLint() {
         var the_symbol = symbol(id, bp);
         the_symbol.led = function (left) {
             var the_token = token;
-            the_token.arity = "binary";
+            the_token.arity = 'binary';
             if (f !== undefined) {
                 return f(left);
             }
@@ -2222,7 +2187,7 @@ var jslint = (function JSLint() {
         var the_symbol = symbol(id, 150);
         the_symbol.led = function (left) {
             token.expression = left;
-            token.arity = "post";
+            token.arity = 'post';
             mutation_check(token.expression);
             return token;
         };
@@ -2236,7 +2201,7 @@ var jslint = (function JSLint() {
         var the_symbol = symbol(id);
         the_symbol.nud = function () {
             var the_token = token;
-            the_token.arity = "pre";
+            the_token.arity = 'pre';
             the_token.expression = expression(150);
             mutation_check(the_token.expression);
             return the_token;
@@ -2251,8 +2216,8 @@ var jslint = (function JSLint() {
         var the_symbol = symbol(id);
         the_symbol.nud = function () {
             var the_token = token;
-            the_token.arity = "unary";
-            if (typeof f === "function") {
+            the_token.arity = 'unary';
+            if (typeof f === 'function') {
                 return f();
             }
             the_token.expression = expression(150);
@@ -2267,7 +2232,7 @@ var jslint = (function JSLint() {
 
         var the_symbol = symbol(id);
         the_symbol.fud = function () {
-            token.arity = "statement";
+            token.arity = 'statement';
             return f();
         };
         return the_symbol;
@@ -2279,11 +2244,11 @@ var jslint = (function JSLint() {
 
         var the_symbol = symbol(id1, 30);
         the_symbol.led = function (left) {
-            var the_token = token;
-            var second = expression(20);
+            var the_token = token,
+                second = expression(20);
             advance(id2);
-            token.arity = "ternary";
-            the_token.arity = "ternary";
+            token.arity = 'ternary';
+            the_token.arity = 'ternary';
             the_token.expression = [left, second, expression(10)];
             return the_token;
         };
@@ -2294,148 +2259,180 @@ var jslint = (function JSLint() {
 
     syntax = empty();
 
-    symbol("}");
-    symbol(")");
-    symbol("]");
-    symbol(",");
-    symbol(";");
-    symbol(":");
-    symbol("*/");
-    symbol("await");
-    symbol("case");
-    symbol("catch");
-    symbol("class");
-    symbol("default");
-    symbol("else");
-    symbol("enum");
-    symbol("finally");
-    symbol("implements");
-    symbol("interface");
-    symbol("package");
-    symbol("private");
-    symbol("protected");
-    symbol("public");
-    symbol("static");
-    symbol("super");
-    symbol("void");
-    symbol("yield");
+    symbol('}');
+    symbol(')');
+    symbol(']');
+    symbol(',');
+    symbol(';');
+    symbol(':');
+    symbol('*/');
+    symbol('await');
+    symbol('case');
+    symbol('catch');
+    symbol('class');
+    symbol('default');
+    symbol('else');
+    symbol('enum');
+    symbol('finally');
+    symbol('implements');
+    symbol('interface');
+    symbol('package');
+    symbol('private');
+    symbol('protected');
+    symbol('public');
+    symbol('static');
+    symbol('super');
+    symbol('void');
+    symbol('yield');
 
-    constant("(number)", "number");
-    constant("(regexp)", "regexp");
-    constant("(string)", "string");
-    constant("arguments", "object", function () {
-        if (option.es6) {
-            warn("unexpected_a", token);
-        }
+    constant('(number)', 'number');
+    constant('(regexp)', 'regexp');
+    constant('(string)', 'string');
+    constant('arguments', 'object', function () {
+        warn('unexpected_a', token);
         return token;
     });
-    constant("eval", "function", function () {
+    constant('eval', 'function', function () {
         if (!option.eval) {
-            warn("unexpected_a", token);
-        } else if (next_token.id !== "(") {
-            warn("expected_a_before_b", next_token, "(", artifact());
+            warn('unexpected_a', token);
+        } else if (next_token.id !== '(') {
+            warn('expected_a_before_b', next_token, '(', artifact());
         }
         return token;
     });
-    constant("false", "boolean", false);
-    constant("Function", "function", function () {
-        if (!option.eval) {
-            warn("unexpected_a", token);
-        } else if (next_token.id !== "(") {
-            warn("expected_a_before_b", next_token, "(", artifact());
-        }
+    constant('false', 'boolean', false);
+    constant('ignore', 'undefined', function () {
+        warn('unexpected_a', token);
         return token;
     });
-    constant("ignore", "undefined", function () {
-        warn("unexpected_a", token);
-        return token;
-    });
-    constant("Infinity", "number", Infinity);
-    constant("NaN", "number", NaN);
-    constant("null", "null", null);
-    constant("this", "object", function () {
+    constant('Infinity', 'number', Infinity);
+    constant('NaN', 'number', NaN);
+    constant('null', 'null', null);
+    constant('this', 'object', function () {
         if (!option.this) {
-            warn("unexpected_a", token);
+            warn('unexpected_a', token);
         }
         return token;
     });
-    constant("true", "boolean", true);
-    constant("undefined", "undefined");
+    constant('true', 'boolean', true);
+    constant('undefined', 'undefined');
 
-    assignment("=");
-    assignment("+=");
-    assignment("-=");
-    assignment("*=");
-    assignment("/=");
-    assignment("%=");
-    assignment("&=");
-    assignment("|=");
-    assignment("^=");
-    assignment("<<=");
-    assignment(">>=");
-    assignment(">>>=");
+    assignment('=');
+    assignment('+=');
+    assignment('-=');
+    assignment('*=');
+    assignment('/=');
+    assignment('%=');
+    assignment('&=');
+    assignment('|=');
+    assignment('^=');
+    assignment('<<=');
+    assignment('>>=');
+    assignment('>>>=');
 
-    infix("||", 40);
-    infix("&&", 50);
-    infix("|", 70);
-    infix("^", 80);
-    infix("&", 90);
-    infix("==", 100);
-    infix("===", 100);
-    infix("!=", 100);
-    infix("!==", 100);
-    infix("<", 110);
-    infix(">", 110);
-    infix("<=", 110);
-    infix(">=", 110);
-    infix("in", 110);
-    infix("instanceof", 110);
-    infix("<<", 120);
-    infix(">>", 120);
-    infix(">>>", 120);
-    infix("+", 130);
-    infix("-", 130);
-    infix("*", 140);
-    infix("/", 140);
-    infix("%", 140);
-    infix("(", 160, function (left) {
-        var the_paren = token;
-        var the_argument;
-        if (left.id !== "function") {
+    infix('||', 40);
+    infix('&&', 50);
+    infix('|', 70);
+    infix('^', 80);
+    infix('&', 90);
+    infix('==', 100);
+    infix('===', 100);
+    infix('!=', 100);
+    infix('!==', 100);
+    infix('<', 110);
+    infix('>', 110);
+    infix('<=', 110);
+    infix('>=', 110);
+    infix('in', 110);
+    infix('instanceof', 110);
+    infix('<<', 120);
+    infix('>>', 120);
+    infix('>>>', 120);
+    infix('+', 130);
+    infix('-', 130);
+    infix('*', 140);
+    infix('/', 140);
+    infix('%', 140);
+    infix('(', 160, function (left) {
+        var the_paren = token,
+            the_argument;
+        if (left.id !== 'function') {
             left_check(left, the_paren);
         }
-        if (functionage.arity === "statement" && left.identifier) {
-            functionage.name.calls[left.id] = left;
-        }
         the_paren.expression = [left];
-        if (next_token.id !== ")") {
+        if (left.identifier) {
+            if (left.new) {
+                if (
+                    left.id.charAt(0) > 'Z' ||
+                    left.id === 'Boolean' ||
+                    left.id === 'Number' ||
+                    left.id === 'String' ||
+                    (left.id === 'Symbol' && option.es6)
+                ) {
+                    warn('unexpected_a', left, 'new');
+                } else if (left.id === 'Function') {
+                    if (!option.eval) {
+                        warn('unexpected_a', left, 'new Function');
+                    }
+                } else if (left.id === 'Array') {
+                    warn('expected_a_b', left, '[]', 'new Array');
+                } else if (left.id === 'Object') {
+                    warn(
+                        'expected_a_b',
+                        left,
+                        'Object.create(null)',
+                        'new Object'
+                    );
+                }
+            } else {
+                if (
+                    left.id.charAt(0) >= 'A' &&
+                    left.id.charAt(0) <= 'Z' &&
+                    left.id !== 'Boolean' &&
+                    left.id !== 'Number' &&
+                    left.id !== 'String' &&
+                    left.id !== 'Symbol'
+                ) {
+                    warn(
+                        'expected_a_before_b',
+                        left,
+                        'new',
+                        artifact(left)
+                    );
+                }
+                if (functionage.arity === 'statement') {
+                    functionage.name.calls[left.id] = left;
+                }
+            }
+        }
+        if (next_token.id !== ')') {
             (function next() {
                 var ellipsis;
-                if (next_token.id === "...") {
+                if (next_token.id === '...') {
                     if (!option.es6) {
-                        warn("es6");
+                        warn('es6');
                     }
                     ellipsis = true;
-                    advance("...");
+                    advance('...');
                 }
                 the_argument = expression(10);
                 if (ellipsis) {
                     the_argument.ellipsis = true;
                 }
                 the_paren.expression.push(the_argument);
-                if (next_token.id === ",") {
-                    advance(",");
+                if (next_token.id === ',') {
+                    advance(',');
                     return next();
                 }
             }());
         }
-        advance(")", the_paren);
+        advance(')', the_paren);
         if (the_paren.expression.length === 2) {
             the_paren.free = true;
             if (the_argument.wrapped === true) {
-                warn("unexpected_a", the_paren);
+                warn('unexpected_a', the_paren);
             }
-            if (the_argument.id === "(") {
+            if (the_argument.id === '(') {
                 the_argument.wrapped = true;
             }
         } else {
@@ -2443,23 +2440,23 @@ var jslint = (function JSLint() {
         }
         return the_paren;
     });
-    infix(".", 170, function (left) {
-        var the_token = token;
-        var name = next_token;
+    infix('.', 170, function (left) {
+        var the_token = token,
+            name = next_token;
         if (
-            (left.id !== "(string)" || name.id !== "indexOf") &&
-            (left.id !== "[" || (
-                name.id !== "concat" && name.id !== "forEach"
+            (left.id !== '(string)' || name.id !== 'indexOf') &&
+            (left.id !== '[' || (
+                name.id !== 'concat' && name.id !== 'forEach'
             )) &&
-            (left.id !== "+" || name.id !== "slice") &&
-            (left.id !== "(regexp)" || (
-                name.id !== "exec" && name.id !== "test"
+            (left.id !== '+' || name.id !== 'slice') &&
+            (left.id !== '(regexp)' || (
+                name.id !== 'exec' && name.id !== 'test'
             ))
         ) {
             left_check(left, the_token);
         }
         if (!name.identifier) {
-            stop("expected_identifier_a");
+            stop('expected_identifier_a');
         }
         advance();
         survey(name);
@@ -2470,199 +2467,199 @@ var jslint = (function JSLint() {
         the_token.expression = left;
         return the_token;
     });
-    infix("[", 170, function (left) {
-        var the_token = token;
-        var the_subscript = expression(0);
+    infix('[', 170, function (left) {
+        var the_token = token,
+            the_subscript = expression(0);
         if (
-            the_subscript.id === "(string)" &&
+            the_subscript.id === '(string)' &&
             rx_identifier.test(the_subscript.value)
         ) {
-            warn("subscript_a", the_subscript);
+            warn('subscript_a', the_subscript);
             survey(the_subscript);
-        } else if (the_subscript.id === "`") {
-            warn("unexpected_a", the_subscript);
+        } else if (the_subscript.id === '`') {
+            warn('unexpected_a', the_subscript);
         }
         left_check(left, the_token);
         the_token.expression = [left, the_subscript];
-        advance("]");
+        advance(']');
         return the_token;
     });
-    infix("=>", 170, function (left) {
-        return stop("expected_a_before_b", left, "(", artifact(left));
+    infix('=>', 170, function (left) {
+        return stop('expected_a_before_b', left, '(', artifact(left));
     });
 
     function do_tick() {
         var the_tick = token;
         if (!option.es6) {
-            warn("es6", the_tick);
+            warn('es6', the_tick);
         }
         the_tick.value = [];
         the_tick.expression = [];
-        if (next_token.id !== "`") {
+        if (next_token.id !== '`') {
             (function part() {
-                advance("(string)");
+                advance('(string)');
                 the_tick.value.push(token);
-                if (next_token.id === "${") {
-                    advance("${");
+                if (next_token.id === '${') {
+                    advance('${');
                     the_tick.expression.push(expression(0));
-                    advance("}");
+                    advance('}');
                     return part();
                 }
             }());
         }
-        advance("`");
+        advance('`');
         return the_tick;
     }
 
-    infix("`", 160, function (left) {
+    infix('`', 160, function (left) {
         var the_tick = do_tick();
         left_check(left, the_tick);
         the_tick.expression = [left].concat(the_tick.expression);
         return the_tick;
     });
 
-    post("++");
-    post("--");
-    pre("++");
-    pre("--");
+    post('++');
+    post('--');
+    pre('++');
+    pre('--');
 
-    prefix("+");
-    prefix("-");
-    prefix("~");
-    prefix("!");
-    prefix("!!");
-    prefix("[", function () {
+    prefix('+');
+    prefix('-');
+    prefix('~');
+    prefix('!');
+    prefix('!!');
+    prefix('[', function () {
         var the_token = token;
         the_token.expression = [];
-        if (next_token.id !== "]") {
+        if (next_token.id !== ']') {
             (function next() {
-                var element;
-                var ellipsis = false;
-                if (next_token.id === "...") {
+                var element,
+                    ellipsis = false;
+                if (next_token.id === '...') {
                     ellipsis = true;
                     if (!option.es6) {
-                        warn("es6");
+                        warn('es6');
                     }
-                    advance("...");
+                    advance('...');
                 }
                 element = expression(10);
                 if (ellipsis) {
                     element.ellipsis = true;
                 }
                 the_token.expression.push(element);
-                if (next_token.id === ",") {
-                    advance(",");
+                if (next_token.id === ',') {
+                    advance(',');
                     return next();
                 }
             }());
         }
-        advance("]");
+        advance(']');
         return the_token;
     });
-    prefix("/=", function () {
-        stop("expected_a_b", token, "/\\=", "/=");
+    prefix('/=', function () {
+        stop('expected_a_b', token, '/\\=', '/=');
     });
-    prefix("=>", function () {
-        return stop("expected_a_before_b", token, "()", "=>");
+    prefix('=>', function () {
+        return stop('expected_a_before_b', token, '()', '=>');
     });
-    prefix("new", function () {
+    prefix('new', function () {
         var the_new = token;
-        var right = expression(160);
-        if (next_token.id !== "(") {
-            warn("expected_a_before_b", next_token, "()", artifact(next_token));
+        next_token.new = true;
+        the_new.expression = expression(150);
+        if (the_new.expression.id !== '(') {
+            warn('expected_a_before_b', next_token, '()', artifact(next_token));
         }
-        the_new.expression = right;
         return the_new;
     });
-    prefix("typeof");
-    prefix("void", function () {
+    prefix('typeof');
+    prefix('void', function () {
         var the_void = token;
-        warn("unexpected_a", the_void);
+        warn('unexpected_a', the_void);
         the_void.expression = expression(0);
         return the_void;
     });
 
     function parameter(list, signature) {
-        var ellipsis = false;
-        var param;
-        if (next_token.id === "{") {
+        var ellipsis = false,
+            param;
+        if (next_token.id === '{') {
             if (!option.es6) {
-                warn("es6");
+                warn('es6');
             }
             param = next_token;
             param.names = [];
-            advance("{");
-            signature.push("{");
+            advance('{');
+            signature.push('{');
             (function subparameter() {
                 var subparam = next_token;
                 if (!subparam.identifier) {
-                    return stop("expected_identifier_a");
+                    return stop('expected_identifier_a');
                 }
                 survey(subparam);
                 advance();
                 signature.push(subparam.id);
-                if (next_token.id === ":") {
-                    advance(":");
+                if (next_token.id === ':') {
+                    advance(':');
                     advance();
                     token.label = subparam;
                     subparam = token;
                     if (!subparam.identifier) {
-                        return stop("expected_identifier_a");
+                        return stop('expected_identifier_a');
                     }
                 }
                 param.names.push(subparam);
-                if (next_token.id === ",") {
-                    advance(",");
+                if (next_token.id === ',') {
+                    advance(',');
                     signature.push(", ");
                     return subparameter();
                 }
             }());
             list.push(param);
-            advance("}");
-            signature.push("}");
-            if (next_token.id === ",") {
-                advance(",");
+            advance('}');
+            signature.push('}');
+            if (next_token.id === ',') {
+                advance(',');
                 signature.push(", ");
                 return parameter(list, signature);
             }
-        } else if (next_token.id === "[") {
+        } else if (next_token.id === '[') {
             if (!option.es6) {
-                warn("es6");
+                warn('es6');
             }
             param = next_token;
             param.names = [];
-            advance("[");
+            advance('[');
             signature.push("[]");
             (function subparameter() {
                 var subparam = next_token;
                 if (!subparam.identifier) {
-                    return stop("expected_identifier_a");
+                    return stop('expected_identifier_a');
                 }
                 advance();
                 param.names.push(subparam);
-                if (next_token.id === ",") {
-                    advance(",");
+                if (next_token.id === ',') {
+                    advance(',');
                     return subparameter();
                 }
             }());
             list.push(param);
-            advance("]");
-            if (next_token.id === ",") {
-                advance(",");
+            advance(']');
+            if (next_token.id === ',') {
+                advance(',');
                 signature.push(", ");
                 return parameter(list, signature);
             }
         } else {
-            if (next_token.id === "...") {
+            if (next_token.id === '...') {
                 if (!option.es6) {
-                    warn("es6");
+                    warn('es6');
                 }
                 ellipsis = true;
                 signature.push("...");
-                advance("...");
+                advance('...');
             }
             if (!next_token.identifier) {
-                return stop("expected_identifier_a");
+                return stop('expected_identifier_a');
             }
             param = next_token;
             list.push(param);
@@ -2671,15 +2668,15 @@ var jslint = (function JSLint() {
             if (ellipsis) {
                 param.ellipsis = true;
             } else {
-                if (next_token.id === "=") {
+                if (next_token.id === '=') {
                     if (!option.es6) {
-                        warn("es6");
+                        warn('es6');
                     }
-                    advance("=");
+                    advance('=');
                     param.expression = expression(0);
                 }
-                if (next_token.id === ",") {
-                    advance(",");
+                if (next_token.id === ',') {
+                    advance(',');
                     signature.push(", ");
                     return parameter(list, signature);
                 }
@@ -2688,14 +2685,13 @@ var jslint = (function JSLint() {
     }
 
     function parameter_list() {
-        var list = [];
-        var signature = ["("];
-        if (next_token.id !== ")" && next_token.id !== "(end)") {
+        var list = [], signature = ['('];
+        if (next_token.id !== ')' && next_token.id !== '(end)') {
             parameter(list, signature);
         }
-        advance(")");
-        signature.push(")");
-        return [list, signature.join("")];
+        advance(')');
+        signature.push(')');
+        return [list, signature.join('')];
     }
 
     function do_function(the_function) {
@@ -2705,12 +2701,12 @@ var jslint = (function JSLint() {
 
 // A function statement must have a name that will be in the parent's scope.
 
-            if (the_function.arity === "statement") {
+            if (the_function.arity === 'statement') {
                 if (!next_token.identifier) {
-                    return stop("expected_identifier_a", next_token);
+                    return stop('expected_identifier_a', next_token);
                 }
                 name = next_token;
-                enroll(name, "variable", true);
+                enroll(name, 'variable', true);
                 the_function.name = name;
                 name.init = true;
                 name.calls = empty();
@@ -2732,14 +2728,14 @@ var jslint = (function JSLint() {
         }
         the_function.level = functionage.level + 1;
         if (mega_mode) {
-            warn("unexpected_a", the_function);
+            warn('unexpected_a', the_function);
         }
 
 // Don't make functions in loops. It is inefficient, and it can lead to scoping
 // errors.
 
         if (functionage.loop > 0) {
-            warn("function_in_loop", the_function);
+            warn('function_in_loop', the_function);
         }
 
 // Give the function properties for storing its names and for observing the
@@ -2754,8 +2750,8 @@ var jslint = (function JSLint() {
         stack.push(functionage);
         functions.push(the_function);
         functionage = the_function;
-        if (the_function.arity !== "statement" && name) {
-            enroll(name, "function", true);
+        if (the_function.arity !== 'statement' && name) {
+            enroll(name, 'function', true);
             name.dead = false;
             name.init = true;
             name.used = 1;
@@ -2763,15 +2759,15 @@ var jslint = (function JSLint() {
 
 // Parse the parameter list.
 
-        advance("(");
+        advance('(');
         token.free = false;
-        token.arity = "function";
+        token.arity = 'function';
         var pl = parameter_list();
         functionage.parameters = pl[0];
         functionage.signature = pl[1];
         functionage.parameters.forEach(function enroll_parameter(name) {
             if (name.identifier) {
-                enroll(name, "parameter", false);
+                enroll(name, 'parameter', false);
             } else {
                 name.names.forEach(enroll_parameter);
             }
@@ -2779,15 +2775,12 @@ var jslint = (function JSLint() {
 
 // The function's body is a block.
 
-        the_function.block = block("body");
-        if (
-            the_function.arity === "statement" &&
-            next_token.line === token.line
-        ) {
-            return stop("unexpected_a", next_token);
+        the_function.block = block('body');
+        if (the_function.arity === 'statement' && next_token.line === token.line) {
+            return stop('unexpected_a', next_token);
         }
-        if (next_token.id === "." || next_token.id === "[") {
-            warn("unexpected_a");
+        if (next_token.id === '.' || next_token.id === '[') {
+            warn('unexpected_a');
         }
 
 // Restore the previous context.
@@ -2796,20 +2789,20 @@ var jslint = (function JSLint() {
         return the_function;
     }
 
-    prefix("function", do_function);
+    prefix('function', do_function);
 
     function fart(pl) {
-        if (next_token.id === ";") {
-            stop("wrap_assignment", token);
+        if (next_token.id === ';') {
+            stop('wrap_assignment', token);
         }
-        advance("=>");
+        advance('=>');
         var the_arrow = token;
-        the_arrow.arity = "binary";
+        the_arrow.arity = 'binary';
         the_arrow.name = "=>";
         the_arrow.level = functionage.level + 1;
         functions.push(the_arrow);
         if (functionage.loop > 0) {
-            warn("function_in_loop", the_arrow);
+            warn('function_in_loop', the_arrow);
         }
 
 // Give the function properties storing its names and for observing the depth
@@ -2826,14 +2819,14 @@ var jslint = (function JSLint() {
         the_arrow.parameters = pl[0];
         the_arrow.signature = pl[1];
         the_arrow.parameters.forEach(function (name) {
-            enroll(name, "parameter", true);
+            enroll(name, 'parameter', true);
         });
         if (!option.es6) {
-            warn("es6", the_arrow);
+            warn('es6', the_arrow);
         }
-        if (next_token.id === "{") {
-            warn("expected_a_b", the_arrow, "function", "=>");
-            the_arrow.block = block("body");
+        if (next_token.id === '{') {
+            warn('expected_a_b', the_arrow, "function", "=>");
+            the_arrow.block = block('body');
         } else {
             the_arrow.expression = expression(0);
         }
@@ -2841,18 +2834,18 @@ var jslint = (function JSLint() {
         return the_arrow;
     }
 
-    prefix("(", function () {
-        var the_paren = token;
-        var the_value;
-        var cadet = lookahead().id;
+    prefix('(', function () {
+        var the_paren = token,
+            the_value,
+            cadet = lookahead().id;
 
 // We can distinguish between a parameter list for => and a wrapped expression
 // with one token of lookahead.
 
         if (
-            next_token.id === ")" ||
-            next_token.id === "..." ||
-            (next_token.identifier && (cadet === "," || cadet === "="))
+            next_token.id === ')' ||
+            next_token.id === '...' ||
+            (next_token.identifier && (cadet === ',' || cadet === '='))
         ) {
             the_paren.free = false;
             return fart(parameter_list());
@@ -2860,37 +2853,33 @@ var jslint = (function JSLint() {
         the_paren.free = true;
         the_value = expression(0);
         if (the_value.wrapped === true) {
-            warn("unexpected_a", the_paren);
+            warn('unexpected_a', the_paren);
         }
         the_value.wrapped = true;
-        advance(")", the_paren);
+        advance(')', the_paren);
         if (next_token.id === "=>") {
-            if (the_value.arity !== "variable") {
-                if (the_value.id === "{" || the_value.id === "[") {
-                    warn("expected_a_before_b", the_paren, "function", "(");
-                    return stop("expected_a_b", next_token, "{", "=>");
-                }
-                return stop("expected_identifier_a", the_value);
+            if (the_value.arity !== 'variable') {
+                return stop('expected_identifier_a', the_value);
             }
             the_paren.expression = [the_value];
             return fart([the_paren.expression, "(" + the_value.id + ")"]);
         }
         return the_value;
     });
-    prefix("`", do_tick);
-    prefix("{", function () {
-        var the_brace = token;
-        var seen = empty();
+    prefix('`', do_tick);
+    prefix('{', function () {
+        var the_brace = token,
+            seen = empty();
         the_brace.expression = [];
-        if (next_token.id !== "}") {
+        if (next_token.id !== '}') {
             (function member() {
-                var extra = true;
-                var id;
-                var name = next_token;
-                var value;
+                var extra = true,
+                    id,
+                    name = next_token,
+                    value;
                 advance();
                 if (
-                    (name.id === "get" || name.id === "set") &&
+                    (name.id === 'get' || name.id === 'set') &&
                     next_token.identifier
                 ) {
                     extra = name.id;
@@ -2899,100 +2888,100 @@ var jslint = (function JSLint() {
                 }
                 id = survey(name);
                 if (seen[id] === true) {
-                    warn("duplicate_a", name);
-                } else if (seen[id] === "get" && extra !== "set") {
-                    warn("expected_a_before_b", name, "set", artifact(name));
+                    warn('duplicate_a', name);
+                } else if (seen[id] === 'get' && extra !== 'set') {
+                    warn('expected_a_before_b', name, 'set', artifact(name));
                 }
-                seen[id] = (extra === "get")
-                    ? "get"
+                seen[id] = (extra === 'get')
+                    ? 'get'
                     : true;
                 if (name.identifier) {
                     switch (next_token.id) {
-                    case "}":
-                    case ",":
+                    case '}':
+                    case ',':
                         if (!option.es6) {
-                            warn("es6");
+                            warn('es6');
                         } else if (extra !== true) {
-                            advance(":");
+                            advance(':');
                         }
                         value = expression(Infinity, true);
                         break;
-                    case "(":
-                        if (!option.es6 && typeof extra !== "string") {
-                            warn("es6");
+                    case '(':
+                        if (!option.es6 && typeof extra !== 'string') {
+                            warn('es6');
                         }
                         value = do_function({
-                            arity: "unary",
+                            arity: 'unary',
                             from: name.from,
-                            id: "function",
+                            id: 'function',
                             line: name.line,
                             name: name,
                             thru: name.from
                         }, name);
                         break;
                     default:
-                        advance(":");
+                        advance(':');
                         value = expression(0);
                     }
                     value.label = name;
-                    if (typeof extra === "string") {
+                    if (typeof extra === 'string') {
                         value.extra = extra;
                     }
                     the_brace.expression.push(value);
                 } else {
-                    advance(":");
+                    advance(':');
                     value = expression(0);
                     value.label = name;
                     the_brace.expression.push(value);
                 }
-                if (next_token.id === ",") {
-                    advance(",");
+                if (next_token.id === ',') {
+                    advance(',');
                     return member();
                 }
             }());
         }
-        advance("}");
+        advance('}');
         return the_brace;
     });
 
-    stmt(";", function () {
-        warn("unexpected_a", token);
+    stmt(';', function () {
+        warn('unexpected_a', token);
         return token;
     });
-    stmt("{", function () {
-        warn("naked_block", token);
-        return block("naked");
+    stmt('{', function () {
+        warn('naked_block', token);
+        return block('naked');
     });
-    stmt("break", function () {
-        var the_break = token;
-        var the_label;
+    stmt('break', function () {
+        var the_break = token,
+            the_label;
         if (functionage.loop < 1 && functionage.switch < 1) {
-            warn("unexpected_a", the_break);
+            warn('unexpected_a', the_break);
         }
         the_break.disrupt = true;
         if (next_token.identifier && token.line === next_token.line) {
             the_label = functionage.context[next_token.id];
             if (
                 the_label === undefined ||
-                the_label.role !== "label" ||
+                the_label.role !== 'label' ||
                 the_label.dead
             ) {
                 warn((the_label !== undefined && the_label.dead)
-                    ? "out_of_scope_a"
-                    : "not_label_a");
+                    ? 'out_of_scope_a'
+                    : 'not_label_a');
             } else {
                 the_label.used += 1;
             }
             the_break.label = next_token;
             advance();
         }
-        advance(";");
+        advance(';');
         return the_break;
     });
 
     function do_var() {
-        var the_statement = token;
-        var is_const = the_statement.id === "const";
+        var the_statement = token,
+            is_const = the_statement.id === 'const';
         the_statement.names = [];
 
 // A program may use var or let, but not both, and let and const require
@@ -3000,16 +2989,16 @@ var jslint = (function JSLint() {
 
         if (is_const) {
             if (!option.es6) {
-                warn("es6", the_statement);
+                warn('es6', the_statement);
             }
         } else if (var_mode === undefined) {
             var_mode = the_statement.id;
-            if (!option.es6 && var_mode !== "var") {
-                warn("es6", the_statement);
+            if (!option.es6 && var_mode !== 'var') {
+                warn('es6', the_statement);
             }
         } else if (the_statement.id !== var_mode) {
             warn(
-                "expected_a_b",
+                'expected_a_b',
                 the_statement,
                 var_mode,
                 the_statement.id
@@ -3019,94 +3008,91 @@ var jslint = (function JSLint() {
 // We don't expect to see variables created in switch statements.
 
         if (functionage.switch > 0) {
-            warn("var_switch", the_statement);
+            warn('var_switch', the_statement);
         }
-        if (functionage.loop > 0 && the_statement.id === "var") {
-            warn("var_loop", the_statement);
+        if (functionage.loop > 0 && the_statement.id === 'var') {
+            warn('var_loop', the_statement);
         }
         (function next() {
-            if (next_token.id === "{" && the_statement.id !== "var") {
+            if (next_token.id === '{' && the_statement.id !== 'var') {
                 var the_brace = next_token;
                 the_brace.names = [];
-                advance("{");
+                advance('{');
                 (function pair() {
                     if (!next_token.identifier) {
-                        return stop("expected_identifier_a", next_token);
+                        return stop('expected_identifier_a', next_token);
                     }
                     var name = next_token;
                     survey(name);
                     advance();
-                    if (next_token.id === ":") {
-                        advance(":");
+                    if (next_token.id === ':') {
+                        advance(':');
                         if (!next_token.identifier) {
-                            return stop("expected_identifier_a", next_token);
+                            return stop('expected_identifier_a', next_token);
                         }
                         next_token.label = name;
                         the_brace.names.push(next_token);
-                        enroll(next_token, "variable", is_const);
+                        enroll(next_token, 'variable', is_const);
                         advance();
                     } else {
                         the_brace.names.push(name);
-                        enroll(name, "variable", is_const);
+                        enroll(name, 'variable', is_const);
                     }
-                    if (next_token.id === ",") {
-                        advance(",");
+                    if (next_token.id === ',') {
+                        advance(',');
                         return pair();
                     }
                 }());
-                advance("}");
-                advance("=");
+                advance('}');
+                advance('=');
                 the_brace.expression = expression(0);
                 the_statement.names.push(the_brace);
-            } else if (next_token.id === "[" && the_statement.id !== "var") {
+            } else if (next_token.id === '[' && the_statement.id !== 'var') {
                 var the_bracket = next_token;
                 the_bracket.names = [];
-                advance("[");
+                advance('[');
                 (function element() {
                     var ellipsis;
-                    if (next_token.id === "...") {
+                    if (next_token.id === '...') {
                         ellipsis = true;
-                        advance("...");
+                        advance('...');
                     }
                     if (!next_token.identifier) {
-                        return stop("expected_identifier_a", next_token);
+                        return stop('expected_identifier_a', next_token);
                     }
                     var name = next_token;
                     advance();
                     the_bracket.names.push(name);
-                    enroll(name, "variable", the_statement.id === "const");
+                    enroll(name, 'variable', the_statement.id === 'const');
                     if (ellipsis) {
                         name.ellipsis = true;
-                    } else if (next_token.id === ",") {
-                        advance(",");
+                    } else if (next_token.id === ',') {
+                        advance(',');
                         return element();
                     }
                 }());
-                advance("]");
-                advance("=");
+                advance(']');
+                advance('=');
                 the_bracket.expression = expression(0);
                 the_statement.names.push(the_bracket);
             } else if (next_token.identifier) {
                 var name = next_token;
                 advance();
-                if (name.id === "ignore") {
-                    warn("unexpected_a", name);
+                if (name.id === 'ignore') {
+                    warn('unexpected_a', name);
                 }
-                enroll(name, "variable", is_const);
-                if (next_token.id === "=" || is_const) {
-                    advance("=");
+                enroll(name, 'variable', is_const);
+                if (next_token.id === '=' || is_const) {
+                    advance('=');
                     name.expression = expression(0);
                     name.init = true;
                 }
                 the_statement.names.push(name);
             } else {
-                return stop("expected_identifier_a", next_token);
+                return stop('expected_identifier_a', next_token);
             }
-            if (next_token.id === ",") {
-                if (!option.multivar) {
-                    warn("expected_a_b", next_token, ";", ",");
-                }
-                advance(",");
+            if (next_token.id === ',') {
+                advance(',');
                 return next();
             }
         }());
@@ -3117,335 +3103,325 @@ var jslint = (function JSLint() {
         return the_statement;
     }
 
-    stmt("const", do_var);
-    stmt("continue", function () {
+    stmt('const', do_var);
+    stmt('continue', function () {
         var the_continue = token;
         if (functionage.loop < 1) {
-            warn("unexpected_a", the_continue);
+            warn('unexpected_a', the_continue);
         }
         not_top_level(the_continue);
         the_continue.disrupt = true;
-        warn("unexpected_a", the_continue);
-        advance(";");
+        warn('unexpected_a', the_continue);
+        advance(';');
         return the_continue;
     });
-    stmt("debugger", function () {
+    stmt('debugger', function () {
         var the_debug = token;
         if (!option.devel) {
-            warn("unexpected_a", the_debug);
+            warn('unexpected_a', the_debug);
         }
         semicolon();
         return the_debug;
     });
-    stmt("delete", function () {
-        var the_token = token;
-        var the_value = expression(0);
+    stmt('delete', function () {
+        var the_token = token,
+            the_value = expression(0);
         if (
-            (the_value.id !== "." && the_value.id !== "[") ||
-            the_value.arity !== "binary"
+            (the_value.id !== '.' && the_value.id !== '[') ||
+            the_value.arity !== 'binary'
         ) {
-            stop("expected_a_b", the_value, ".", artifact(the_value));
+            stop('expected_a_b', the_value, '.', artifact(the_value));
         }
         the_token.expression = the_value;
         semicolon();
         return the_token;
     });
-    stmt("do", function () {
+    stmt('do', function () {
         var the_do = token;
         not_top_level(the_do);
         functionage.loop += 1;
         the_do.block = block();
-        advance("while");
+        advance('while');
         the_do.expression = condition();
         semicolon();
         if (the_do.block.disrupt === true) {
-            warn("weird_loop", the_do);
+            warn('weird_loop', the_do);
         }
         functionage.loop -= 1;
         return the_do;
     });
-    stmt("export", function () {
+    stmt('export', function () {
         var the_export = token;
         if (!option.es6) {
-            warn("es6", the_export);
+            warn('es6', the_export);
         }
-        if (typeof module_mode === "object") {
-            warn("unexpected_directive_a", module_mode, module_mode.directive);
+        if (typeof module_mode === 'object') {
+            warn('unexpected_directive_a', module_mode, module_mode.directive);
         }
-        advance("default");
+        advance('default');
         if (export_mode) {
-            warn("duplicate_a", token);
+            warn('duplicate_a', token);
         }
-        global.strict = true;
         module_mode = true;
         export_mode = true;
         the_export.expression = expression(0);
         semicolon();
         return the_export;
     });
-    stmt("for", function () {
-        var first;
-        var the_for = token;
+    stmt('for', function () {
+        var first,
+            the_for = token;
         if (!option.for) {
-            warn("unexpected_a", the_for);
+            warn('unexpected_a', the_for);
         }
         not_top_level(the_for);
         functionage.loop += 1;
-        advance("(");
+        advance('(');
         token.free = true;
-        if (next_token.id === ";") {
-            return stop("expected_a_b", the_for, "while (", "for (;");
+        if (next_token.id === ';') {
+            return stop('expected_a_b', the_for, 'while (', 'for (;');
         }
         if (
-            next_token.id === "var" ||
-            next_token.id === "let" ||
-            next_token.id === "const"
+            next_token.id === 'var' ||
+            next_token.id === 'let' ||
+            next_token.id === 'const'
         ) {
-            return stop("unexpected_a");
+            return stop('unexpected_a');
         }
         first = expression(0);
-        if (first.id === "in") {
-            if (first.expression[0].arity !== "variable") {
-                warn("bad_assignment_a", first.expression[0]);
+        if (first.id === 'in') {
+            if (first.expression[0].arity !== 'variable') {
+                warn('bad_assignment_a', first.expression[0]);
             }
             the_for.name = first.expression[0];
             the_for.expression = first.expression[1];
-            warn("expected_a_b", the_for, "Object.keys", "for in");
+            warn('expected_a_b', the_for, 'Object.keys', 'for in');
         } else {
             the_for.initial = first;
-            advance(";");
+            advance(';');
             the_for.expression = expression(0);
-            advance(";");
+            advance(';');
             the_for.inc = expression(0);
-            if (the_for.inc.id === "++") {
-                warn("expected_a_b", the_for.inc, "+= 1", "++");
+            if (the_for.inc.id === '++') {
+                warn('expected_a_b', the_for.inc, '+= 1', '++');
             }
         }
-        advance(")");
+        advance(')');
         the_for.block = block();
         if (the_for.block.disrupt === true) {
-            warn("weird_loop", the_for);
+            warn('weird_loop', the_for);
         }
         functionage.loop -= 1;
         return the_for;
     });
-    stmt("function", do_function);
-    stmt("if", function () {
-        var the_else;
-        var the_if = token;
+    stmt('function', do_function);
+    stmt('if', function () {
+        var the_else,
+            the_if = token;
         the_if.expression = condition();
         the_if.block = block();
-        if (next_token.id === "else") {
-            advance("else");
+        if (next_token.id === 'else') {
+            advance('else');
             the_else = token;
-            the_if.else = (next_token.id === "if")
+            the_if.else = (next_token.id === 'if')
                 ? statement()
                 : block();
             if (the_if.block.disrupt === true) {
                 if (the_if.else.disrupt === true) {
                     the_if.disrupt = true;
                 } else {
-                    warn("unexpected_a", the_else);
+                    warn('unexpected_a', the_else);
                 }
             }
         }
         return the_if;
     });
-    stmt("import", function () {
+    stmt('import', function () {
         var the_import = token;
         if (!option.es6) {
-            warn("es6", the_import);
-        } else if (typeof module_mode === "object") {
-            warn("unexpected_directive_a", module_mode, module_mode.directive);
+            warn('es6', the_import);
+        } else if (typeof module_mode === 'object') {
+            warn('unexpected_directive_a', module_mode, module_mode.directive);
         }
         module_mode = true;
         if (!next_token.identifier) {
-            return stop("expected_identifier_a");
+            return stop('expected_identifier_a');
         }
         var name = next_token;
         advance();
-        if (name.id === "ignore") {
-            warn("unexpected_a", name);
+        if (name.id === 'ignore') {
+            warn('unexpected_a', name);
         }
-        enroll(name, "variable", true);
-        advance("from");
-        advance("(string)");
+        enroll(name, 'variable', true);
+        advance('from');
+        advance('(string)');
         the_import.import = token;
         the_import.name = name;
         if (!rx_identifier.test(token.value)) {
-            warn("bad_module_name_a", token);
+            warn('bad_module_name_a', token);
         }
         imports.push(token.value);
         semicolon();
         return the_import;
     });
-    stmt("let", do_var);
-    stmt("return", function () {
+    stmt('let', do_var);
+    stmt('return', function () {
         var the_return = token;
         not_top_level(the_return);
         the_return.disrupt = true;
-        if (next_token.id !== ";" && the_return.line === next_token.line) {
+        if (next_token.id !== ';' && the_return.line === next_token.line) {
             the_return.expression = expression(10);
         }
-        advance(";");
+        advance(';');
         return the_return;
     });
-    stmt("switch", function () {
-        var dups = [];
-        var last;
-        var stmts;
-        var the_cases = [];
-        var the_disrupt = true;
-        var the_switch = token;
+    stmt('switch', function () {
+        var dups = [],
+            last,
+            stmts,
+            the_cases = [],
+            the_disrupt = true,
+            the_switch = token;
         not_top_level(the_switch);
         functionage.switch += 1;
-        advance("(");
+        advance('(');
         token.free = true;
         the_switch.expression = expression(0);
         the_switch.block = the_cases;
-        advance(")");
-        advance("{");
+        advance(')');
+        advance('{');
         (function major() {
             var the_case = next_token;
-            the_case.arity = "statement";
+            the_case.arity = 'statement';
             the_case.expression = [];
             (function minor() {
-                advance("case");
+                advance('case');
                 token.switch = true;
                 var exp = expression(0);
                 if (dups.some(function (thing) {
                     return are_similar(thing, exp);
                 })) {
-                    warn("unexpected_a", exp);
+                    warn('unexpected_a', exp);
                 }
                 dups.push(exp);
                 the_case.expression.push(exp);
-                advance(":");
-                if (next_token.id === "case") {
+                advance(':');
+                if (next_token.id === 'case') {
                     return minor();
                 }
             }());
             stmts = statements();
             if (stmts.length < 1) {
-                warn("expected_statements_a");
+                warn('expected_statements_a');
                 return;
             }
             the_case.block = stmts;
             the_cases.push(the_case);
             last = stmts[stmts.length - 1];
             if (last.disrupt) {
-                if (last.id === "break" && last.label === undefined) {
+                if (last.id === 'break' && last.label === undefined) {
                     the_disrupt = false;
                 }
             } else {
                 warn(
-                    "expected_a_before_b",
+                    'expected_a_before_b',
                     next_token,
-                    "break;",
+                    'break;',
                     artifact(next_token)
                 );
             }
-            if (next_token.id === "case") {
+            if (next_token.id === 'case') {
                 return major();
             }
         }());
         dups = undefined;
-        if (next_token.id === "default") {
-            var the_default = next_token;
-            advance("default");
+        if (next_token.id === 'default') {
+            advance('default');
             token.switch = true;
-            advance(":");
+            advance(':');
             the_switch.else = statements();
             if (the_switch.else.length < 1) {
-                warn("unexpected_a", the_default);
+                warn('expected_statements_a');
                 the_disrupt = false;
             } else {
-                var the_last = the_switch.else[the_switch.else.length - 1];
-                if (the_last.id === "break" && the_last.label === undefined) {
-                    warn("unexpected_a", the_last);
-                    the_last.disrupt = false;
-                }
-                the_disrupt = the_disrupt && the_last.disrupt;
+                the_disrupt =
+                        the_disrupt &&
+                        the_switch.else[the_switch.else.length - 1].disrupt;
             }
         } else {
             the_disrupt = false;
         }
-        advance("}", the_switch);
+        advance('}', the_switch);
         functionage.switch -= 1;
         the_switch.disrupt = the_disrupt;
         return the_switch;
     });
-    stmt("throw", function () {
+    stmt('throw', function () {
         var the_throw = token;
         the_throw.disrupt = true;
         the_throw.expression = expression(10);
         semicolon();
         return the_throw;
     });
-    stmt("try", function () {
-        var clause = false;
-        var the_catch;
-        var the_disrupt;
-        var the_try = token;
+    stmt('try', function () {
+        var clause = false,
+            the_catch,
+            the_disrupt,
+            the_try = token;
         the_try.block = block();
         the_disrupt = the_try.block.disrupt;
-        if (next_token.id === "catch") {
-            var ignored = "ignore";
+        if (next_token.id === 'catch') {
+            var ignored = 'ignore';
             clause = true;
             the_catch = next_token;
             the_try.catch = the_catch;
-            advance("catch");
-            advance("(");
+            advance('catch');
+            advance('(');
             if (!next_token.identifier) {
-                return stop("expected_identifier_a", next_token);
+                return stop('expected_identifier_a', next_token);
             }
-            if (next_token.id !== "ignore") {
+            if (next_token.id !== 'ignore') {
                 ignored = undefined;
                 the_catch.name = next_token;
-                enroll(next_token, "exception", true);
+                enroll(next_token, 'exception', true);
             }
             advance();
-            advance(")");
+            advance(')');
             the_catch.block = block(ignored);
             if (the_catch.block.disrupt !== true) {
                 the_disrupt = false;
             }
         }
-        if (next_token.id === "finally") {
+        if (next_token.id === 'finally') {
             clause = true;
-            advance("finally");
+            advance('finally');
             the_try.else = block();
             the_disrupt = the_try.else.disrupt;
         }
         the_try.disrupt = the_disrupt;
         if (!clause) {
-            warn(
-                "expected_a_before_b",
-                next_token,
-                "catch",
-                artifact(next_token)
-            );
+            warn('expected_a_before_b', next_token, 'catch', artifact(next_token));
         }
         return the_try;
     });
-    stmt("var", do_var);
-    stmt("while", function () {
+    stmt('var', do_var);
+    stmt('while', function () {
         var the_while = token;
         not_top_level(the_while);
         functionage.loop += 1;
         the_while.expression = condition();
         the_while.block = block();
         if (the_while.block.disrupt === true) {
-            warn("weird_loop", the_while);
+            warn('weird_loop', the_while);
         }
         functionage.loop -= 1;
         return the_while;
     });
-    stmt("with", function () {
-        stop("unexpected_a", token);
+    stmt('with', function () {
+        stop('unexpected_a', token);
     });
 
-    ternary("?", ":");
+    ternary('?', ':');
 
 // Ambulation of the parse tree.
 
@@ -3455,15 +3431,15 @@ var jslint = (function JSLint() {
 // the tree is traversed.
 
         return function (arity, id, task) {
-            var a_set = when[arity];
-            var i_set;
+            var a_set = when[arity],
+                i_set;
 
 // The id parameter is optional. If excluded, the task will be applied to all
 // ids.
 
-            if (typeof id !== "string") {
+            if (typeof id !== 'string') {
                 task = id;
-                id = "(all)";
+                id = '(all)';
             }
 
 // If this arity has no registrations yet, then create a set object to hold
@@ -3498,8 +3474,8 @@ var jslint = (function JSLint() {
 // Given a task set that was built by an action function, run all of the
 // relevant tasks on the token.
 
-            var a_set = when[the_token.arity];
-            var i_set;
+            var a_set = when[the_token.arity],
+                i_set;
 
 // If there are tasks associated with the token's arity...
 
@@ -3516,7 +3492,7 @@ var jslint = (function JSLint() {
 
 // If there are tasks for all ids.
 
-                i_set = a_set["(all)"];
+                i_set = a_set['(all)'];
                 if (i_set !== undefined) {
                     i_set.forEach(function (task) {
                         return task(the_token);
@@ -3526,12 +3502,12 @@ var jslint = (function JSLint() {
         };
     }
 
-    var posts = empty();
-    var pres = empty();
-    var preaction = action(pres);
-    var postaction = action(posts);
-    var preamble = amble(pres);
-    var postamble = amble(posts);
+    var posts = empty(),
+        pres = empty(),
+        preaction = action(pres),
+        postaction = action(posts),
+        preamble = amble(pres),
+        postamble = amble(posts);
 
     function walk_expression(thing) {
         if (thing) {
@@ -3540,17 +3516,17 @@ var jslint = (function JSLint() {
             } else {
                 preamble(thing);
                 walk_expression(thing.expression);
-                if (thing.id === "function") {
+                if (thing.id === 'function') {
                     walk_statement(thing.block);
                 }
                 switch (thing.arity) {
-                case "post":
-                case "pre":
-                    warn("unexpected_a", thing);
+                case 'post':
+                case 'pre':
+                    warn('unexpected_a', thing);
                     break;
-                case "statement":
-                case "assignment":
-                    warn("unexpected_statement_a", thing);
+                case 'statement':
+                case 'assignment':
+                    warn('unexpected_statement_a', thing);
                     break;
                 }
                 postamble(thing);
@@ -3566,16 +3542,16 @@ var jslint = (function JSLint() {
                 preamble(thing);
                 walk_expression(thing.expression);
                 switch (thing.arity) {
-                case "statement":
-                case "assignment":
+                case 'statement':
+                case 'assignment':
                     break;
-                case "binary":
-                    if (thing.id !== "(") {
-                        warn("unexpected_expression_a", thing);
+                case 'binary':
+                    if (thing.id !== '(') {
+                        warn('unexpected_expression_a', thing);
                     }
                     break;
                 default:
-                    warn("unexpected_expression_a", thing);
+                    warn('unexpected_expression_a', thing);
                 }
                 walk_statement(thing.block);
                 walk_statement(thing.else);
@@ -3585,7 +3561,7 @@ var jslint = (function JSLint() {
     }
 
     function lookup(thing) {
-        if (thing.arity === "variable") {
+        if (thing.arity === 'variable') {
 
 // Look up the variable in the current context.
 
@@ -3599,7 +3575,7 @@ var jslint = (function JSLint() {
                     var a_variable = outer.context[thing.id];
                     if (
                         a_variable !== undefined &&
-                        a_variable.role !== "label"
+                        a_variable.role !== 'label'
                     ) {
                         the_variable = a_variable;
                     }
@@ -3610,7 +3586,7 @@ var jslint = (function JSLint() {
 
                 if (the_variable === undefined) {
                     if (declared_globals[thing.id] === undefined) {
-                        warn("undeclared_a", thing);
+                        warn('undeclared_a', thing);
                         return;
                     }
                     the_variable = {
@@ -3618,7 +3594,7 @@ var jslint = (function JSLint() {
                         function: global,
                         id: thing.id,
                         init: true,
-                        role: "variable",
+                        role: 'variable',
                         used: 0,
                         writable: false
                     };
@@ -3626,11 +3602,11 @@ var jslint = (function JSLint() {
                 }
                 the_variable.closure = true;
                 functionage.context[thing.id] = the_variable;
-            } else if (the_variable.role === "label") {
-                warn("label_a", thing);
+            } else if (the_variable.role === 'label') {
+                warn('label_a', thing);
             }
             if (the_variable.dead) {
-                warn("out_of_scope_a", thing);
+                warn('out_of_scope_a', thing);
             }
             return the_variable;
         }
@@ -3643,33 +3619,33 @@ var jslint = (function JSLint() {
     }
 
     function preaction_function(thing) {
-        if (thing.arity === "statement" && blockage.body !== true) {
-            warn("unexpected_a", thing);
+        if (thing.arity === 'statement' && blockage.body !== true) {
+            warn('unexpected_a', thing);
         }
         stack.push(functionage);
         block_stack.push(blockage);
         functionage = thing;
         blockage = thing;
         thing.live = [];
-        if (typeof thing.name === "object") {
+        if (typeof thing.name === 'object') {
             thing.name.dead = false;
             thing.name.init = true;
         }
         switch (thing.extra) {
-        case "get":
+        case 'get':
             if (thing.parameters.length !== 0) {
-                warn("bad_get", thing);
+                warn('bad_get', thing);
             }
             break;
-        case "set":
+        case 'set':
             if (thing.parameters.length !== 1) {
-                warn("bad_set", thing);
+                warn('bad_set', thing);
             }
             break;
         }
         thing.parameters.forEach(function (name) {
             walk_expression(name.expression);
-            if (name.id === "{" || name.id === "[") {
+            if (name.id === '{' || name.id === '[') {
                 name.names.forEach(subactivate);
             } else {
                 name.dead = false;
@@ -3680,20 +3656,20 @@ var jslint = (function JSLint() {
 
     function bitwise_check(thing) {
         if (!option.bitwise && bitwiseop[thing.id] === true) {
-            warn("unexpected_a", thing);
+            warn('unexpected_a', thing);
         }
         if (
-            thing.id !== "(" &&
-            thing.id !== "&&" &&
-            thing.id !== "||" &&
-            thing.id !== "=" &&
+            thing.id !== '(' &&
+            thing.id !== '&&' &&
+            thing.id !== '||' &&
+            thing.id !== '=' &&
             Array.isArray(thing.expression) &&
             thing.expression.length === 2 && (
                 relationop[thing.expression[0].id] === true ||
                 relationop[thing.expression[1].id] === true
             )
         ) {
-            warn("unexpected_a", thing);
+            warn('unexpected_a', thing);
         }
     }
 
@@ -3708,7 +3684,7 @@ var jslint = (function JSLint() {
     function activate(name) {
         if (name.expression !== undefined) {
             walk_expression(name.expression);
-            if (name.id === "{" || name.id === "[") {
+            if (name.id === '{' || name.id === '[') {
                 name.names.forEach(subactivate);
             } else {
                 name.init = true;
@@ -3722,60 +3698,60 @@ var jslint = (function JSLint() {
         thing.names.forEach(activate);
     }
 
-    preaction("assignment", bitwise_check);
-    preaction("binary", bitwise_check);
-    preaction("binary", function (thing) {
+    preaction('assignment', bitwise_check);
+    preaction('binary', bitwise_check);
+    preaction('binary', function (thing) {
         if (relationop[thing.id] === true) {
-            var left = thing.expression[0];
-            var right = thing.expression[1];
-            if (left.id === "NaN" || right.id === "NaN") {
-                warn("isNaN", thing);
-            } else if (left.id === "typeof") {
-                if (right.id !== "(string)") {
-                    if (right.id !== "typeof") {
-                        warn("expected_string_a", right);
+            var left = thing.expression[0],
+                right = thing.expression[1];
+            if (left.id === 'NaN' || right.id === 'NaN') {
+                warn('isNaN', thing);
+            } else if (left.id === 'typeof') {
+                if (right.id !== '(string)') {
+                    if (right.id !== 'typeof') {
+                        warn('expected_string_a', right);
                     }
                 } else {
                     var value = right.value;
-                    if (value === "symbol") {
+                    if (value === 'symbol') {
                         if (!option.es6) {
-                            warn("es6", right, value);
+                            warn('es6', right, value);
                         }
-                    } else if (value === "null" || value === "undefined") {
-                        warn("unexpected_typeof_a", right, value);
+                    } else if (value === 'null' || value === 'undefined') {
+                        warn('unexpected_typeof_a', right, value);
                     } else if (
-                        value !== "boolean" &&
-                        value !== "function" &&
-                        value !== "number" &&
-                        value !== "object" &&
-                        value !== "string"
+                        value !== 'boolean' &&
+                        value !== 'function' &&
+                        value !== 'number' &&
+                        value !== 'object' &&
+                        value !== 'string'
                     ) {
-                        warn("expected_type_string_a", right, value);
+                        warn('expected_type_string_a', right, value);
                     }
                 }
             }
         }
     });
-    preaction("binary", "==", function (thing) {
-        warn("expected_a_b", thing, "===", "==");
+    preaction('binary', '==', function (thing) {
+        warn('expected_a_b', thing, '===', '==');
     });
-    preaction("binary", "!=", function (thing) {
-        warn("expected_a_b", thing, "!==", "!=");
+    preaction('binary', '!=', function (thing) {
+        warn('expected_a_b', thing, '!==', '!=');
     });
-    preaction("binary", "=>", preaction_function);
-    preaction("binary", "||", function (thing) {
+    preaction('binary', '=>', preaction_function);
+    preaction('binary', '||', function (thing) {
         thing.expression.forEach(function (thang) {
-            if (thang.id === "&&" && !thang.wrapped) {
-                warn("and", thang);
+            if (thang.id === '&&' && !thang.wrapped) {
+                warn('and', thang);
             }
         });
     });
-    preaction("binary", "(", function (thing) {
+    preaction('binary', '(', function (thing) {
         var left = thing.expression[0];
         if (
             left.identifier &&
             functionage.context[left.id] === undefined &&
-            typeof functionage.name === "object"
+            typeof functionage.name === 'object'
         ) {
             var parent = functionage.name.function;
             if (parent) {
@@ -3792,38 +3768,30 @@ var jslint = (function JSLint() {
             }
         }
     });
-    preaction("binary", "in", function (thing) {
-        warn("infix_in", thing);
+    preaction('binary', 'in', function (thing) {
+        warn('infix_in', thing);
     });
-    preaction("binary", "instanceof", function (thing) {
-        warn("unexpected_a", thing);
-    });
-    preaction("binary", ".", function (thing) {
-        if (thing.expression.new) {
-            thing.new = true;
-        }
-    });
-    preaction("statement", "{", function (thing) {
+    preaction('statement', '{', function (thing) {
         block_stack.push(blockage);
         blockage = thing;
         thing.live = [];
     });
-    preaction("statement", "for", function (thing) {
+    preaction('statement', 'for', function (thing) {
         if (thing.name !== undefined) {
             var the_variable = lookup(thing.name);
             if (the_variable !== undefined) {
                 the_variable.init = true;
                 if (!the_variable.writable) {
-                    warn("bad_assignment_a", thing.name);
+                    warn('bad_assignment_a', thing.name);
                 }
             }
         }
         walk_statement(thing.initial);
     });
-    preaction("statement", "function", preaction_function);
-    preaction("unary", "~", bitwise_check);
-    preaction("unary", "function", preaction_function);
-    preaction("variable", function (thing) {
+    preaction('statement', 'function', preaction_function);
+    preaction('unary', '~', bitwise_check);
+    preaction('unary', 'function', preaction_function);
+    preaction('variable', function (thing) {
         var the_variable = lookup(thing);
         if (the_variable !== undefined) {
             thing.variable = the_variable;
@@ -3839,56 +3807,29 @@ var jslint = (function JSLint() {
                 return;
             }
         }
-        warn("bad_assignment_a", name);
+        warn('bad_assignment_a', name);
     }
 
-    postaction("assignment", function (thing) {
+    postaction('assignment', function (thing) {
 
 // Assignment using = sets the init property of a variable. No other assignment
 // operator can do this. A = token keeps that variable (or array of variables
 // in case of destructuring) in its name property.
 
-        var lvalue = thing.expression[0];
-        if (thing.id === "=") {
+        if (thing.id === '=') {
             if (thing.names !== undefined) {
                 if (Array.isArray(thing.names)) {
                     thing.names.forEach(init_variable);
                 } else {
                     init_variable(thing.names);
                 }
-            } else {
-                if (
-                    lvalue.id === "." &&
-                    thing.expression[1].id === "undefined"
-                ) {
-                    warn(
-                        "expected_a_b",
-                        lvalue.expression,
-                        "delete",
-                        "undefined"
-                    );
-                }
             }
         } else {
-            if (lvalue.arity === "variable") {
+            var lvalue = thing.expression[0];
+            if (lvalue.arity === 'variable') {
                 if (!lvalue.variable || lvalue.variable.writable !== true) {
-                    warn("bad_assignment_a", lvalue);
+                    warn('bad_assignment_a', lvalue);
                 }
-            }
-            var right = syntax[thing.expression[1].id];
-            if (
-                right !== undefined &&
-                (
-                    right.id === "function" ||
-                    right.id === "=>" ||
-                    (
-                        right.constant &&
-                        right.id !== "(number)" &&
-                        (right.id !== "(string)" || thing.id !== "+=")
-                    )
-                )
-            ) {
-                warn("unexpected_a", thing.expression[1]);
             }
         }
     });
@@ -3898,15 +3839,12 @@ var jslint = (function JSLint() {
         delete functionage.switch;
         functionage = stack.pop();
         if (thing.wrapped) {
-            warn("unexpected_parens", thing);
-        }
-        if (typeof thing.name === "object") {
-            thing.name.used = 0;
+            warn('unexpected_parens', thing);
         }
         return pop_block();
     }
 
-    postaction("binary", function (thing) {
+    postaction('binary', function (thing) {
         var right;
         if (relationop[thing.id]) {
             if (
@@ -3918,27 +3856,27 @@ var jslint = (function JSLint() {
                     thing.expression[1].constant === true
                 )
             ) {
-                warn("weird_relation_a", thing);
+                warn('weird_relation_a', thing);
             }
         }
         switch (thing.id) {
-        case "+":
-        case "-":
+        case '+':
+        case '-':
             right = thing.expression[1];
             if (
                 right.id === thing.id &&
-                right.arity === "unary" &&
+                right.arity === 'unary' &&
                 !right.wrapped
             ) {
-                warn("wrap_unary", right);
+                warn('wrap_unary', right);
             }
             break;
-        case "=>":
-        case "(":
+        case '=>':
+        case '(':
             break;
-        case ".":
-            if (thing.expression.id === "RegExp") {
-                warn("weird_expression_a", thing);
+        case '.':
+            if (thing.expression.id === 'RegExp') {
+                warn('weird_expression_a', thing);
             }
             break;
         default:
@@ -3950,138 +3888,55 @@ var jslint = (function JSLint() {
             }
         }
     });
-    postaction("binary", "&&", function (thing) {
+    postaction('binary', '&&', function (thing) {
         if (
             is_weird(thing.expression[0]) ||
             are_similar(thing.expression[0], thing.expression[1]) ||
             thing.expression[0].constant === true ||
             thing.expression[1].constant === true
         ) {
-            warn("weird_condition_a", thing);
+            warn('weird_condition_a', thing);
         }
     });
-    postaction("binary", "||", function (thing) {
+    postaction('binary', '||', function (thing) {
         if (
             is_weird(thing.expression[0]) ||
             are_similar(thing.expression[0], thing.expression[1]) ||
             thing.expression[0].constant === true
         ) {
-            warn("weird_condition_a", thing);
+            warn('weird_condition_a', thing);
         }
     });
-    postaction("binary", "=>", postaction_function);
-    postaction("binary", "(", function (thing) {
-        var left = thing.expression[0];
-        var the_new;
-        if (left.id === "new") {
-            the_new = left;
-            left = left.expression;
-        }
-        if (left.id === "function") {
-            if (!thing.wrapped) {
-                warn("wrap_immediate", thing);
-            }
-        } else if (left.identifier) {
-            if (the_new !== undefined) {
-                if (
-                    left.id.charAt(0) > "Z" ||
-                    left.id === "Boolean" ||
-                    left.id === "Number" ||
-                    left.id === "String" ||
-                    (left.id === "Symbol" && option.es6)
-                ) {
-                    warn("unexpected_a", the_new);
-                } else if (left.id === "Function") {
-                    if (!option.eval) {
-                        warn("unexpected_a", left, "new Function");
-                    }
-                } else if (left.id === "Array") {
-                    warn("expected_a_b", left, "[]", "new Array");
-                } else if (left.id === "Object") {
-                    warn(
-                        "expected_a_b",
-                        left,
-                        "Object.create(null)",
-                        "new Object"
-                    );
-                }
-            } else {
-                if (
-                    left.id.charAt(0) >= "A" &&
-                    left.id.charAt(0) <= "Z" &&
-                    left.id !== "Boolean" &&
-                    left.id !== "Number" &&
-                    left.id !== "String" &&
-                    left.id !== "Symbol"
-                ) {
-                    warn(
-                        "expected_a_before_b",
-                        left,
-                        "new",
-                        artifact(left)
-                    );
-                }
-            }
-        } else if (left.id === ".") {
-            var cack = the_new !== undefined;
-            if (left.expression.id === "Date" && left.name.id === "UTC") {
-                cack = !cack;
-            }
-            if (rx_cap.test(left.name.id) !== cack) {
-                if (the_new !== undefined) {
-                    warn("unexpected_a", the_new);
-                } else {
-                    warn(
-                        "expected_a_before_b",
-                        left.expression,
-                        "new",
-                        left.name.id
-                    );
-                }
-            }
-            if (left.name.id === "getTime") {
-                var l1 = left.expression;
-                if (l1.id === "(") {
-                    var l2 = l1.expression;
-                    if (l2.length === 1) {
-                        var l3 = l2[0];
-                        if (l3.id === "new" && l3.expression.id === "Date") {
-                            warn(
-                                "expected_a_b",
-                                l3,
-                                "Date.now()",
-                                "new Date().getTime()"
-                            );
-                        }
-                    }
-                }
-            }
+    postaction('binary', '=>', postaction_function);
+    postaction('binary', '(', function (thing) {
+        if (!thing.wrapped && thing.expression[0].id === 'function') {
+            warn('wrap_immediate', thing);
         }
     });
-    postaction("binary", "[", function (thing) {
-        if (thing.expression[0].id === "RegExp") {
-            warn("weird_expression_a", thing);
+    postaction('binary', '[', function (thing) {
+        if (thing.expression[0].id === 'RegExp') {
+            warn('weird_expression_a', thing);
         }
         if (is_weird(thing.expression[1])) {
-            warn("weird_expression_a", thing.expression[1]);
+            warn('weird_expression_a', thing.expression[1]);
         }
     });
-    postaction("statement", "{", pop_block);
-    postaction("statement", "const", action_var);
-    postaction("statement", "export", top_level_only);
-    postaction("statement", "for", function (thing) {
+    postaction('statement', '{', pop_block);
+    postaction('statement', 'const', action_var);
+    postaction('statement', 'export', top_level_only);
+    postaction('statement', 'for', function (thing) {
         walk_statement(thing.inc);
     });
-    postaction("statement", "function", postaction_function);
-    postaction("statement", "import", function (the_thing) {
+    postaction('statement', 'function', postaction_function);
+    postaction('statement', 'import', function (the_thing) {
         var name = the_thing.name;
         name.init = true;
         name.dead = false;
         blockage.live.push(name);
         return top_level_only(the_thing);
     });
-    postaction("statement", "let", action_var);
-    postaction("statement", "try", function (thing) {
+    postaction('statement', 'let', action_var);
+    postaction('statement', 'try', function (thing) {
         if (thing.catch !== undefined) {
             var the_name = thing.catch.name;
             if (the_name !== undefined) {
@@ -4092,43 +3947,43 @@ var jslint = (function JSLint() {
             walk_statement(thing.catch.block);
         }
     });
-    postaction("statement", "var", action_var);
-    postaction("ternary", function (thing) {
+    postaction('statement', 'var', action_var);
+    postaction('ternary', function (thing) {
         if (
             is_weird(thing.expression[0]) ||
             thing.expression[0].constant === true ||
             are_similar(thing.expression[1], thing.expression[2])
         ) {
-            warn("unexpected_a", thing);
+            warn('unexpected_a', thing);
         } else if (are_similar(thing.expression[0], thing.expression[1])) {
-            warn("expected_a_b", thing, "||", "?");
+            warn('expected_a_b', thing, '||', '?');
         } else if (are_similar(thing.expression[0], thing.expression[2])) {
-            warn("expected_a_b", thing, "&&", "?");
+            warn('expected_a_b', thing, '&&', '?');
         } else if (
-            thing.expression[1].id === "true" &&
-            thing.expression[2].id === "false"
+            thing.expression[1].id === 'true' &&
+            thing.expression[2].id === 'false'
         ) {
-            warn("expected_a_b", thing, "!!", "?");
+            warn('expected_a_b', thing, '!!', '?');
         } else if (
-            thing.expression[1].id === "false" &&
-            thing.expression[2].id === "true"
+            thing.expression[1].id === 'false' &&
+            thing.expression[2].id === 'true'
         ) {
-            warn("expected_a_b", thing, "!", "?");
+            warn('expected_a_b', thing, '!', '?');
         } else if (thing.expression[0].wrapped !== true && (
-            thing.expression[0].id === "||" ||
-            thing.expression[0].id === "&&"
+            thing.expression[0].id === '||' ||
+            thing.expression[0].id === '&&'
         )) {
-            warn("wrap_condition", thing.expression[0]);
+            warn('wrap_condition', thing.expression[0]);
         }
     });
-    postaction("unary", function (thing) {
+    postaction('unary', function (thing) {
         switch (thing.id) {
-        case "[":
-        case "{":
-        case "function":
-        case "new":
+        case '[':
+        case '{':
+        case 'function':
+        case 'new':
             break;
-        case "`":
+        case '`':
             if (thing.expression.every(function (thing) {
                 return thing.constant;
             })) {
@@ -4141,20 +3996,17 @@ var jslint = (function JSLint() {
             }
         }
     });
-    postaction("unary", "function", postaction_function);
+    postaction('unary', 'function', postaction_function);
 
     function delve(the_function) {
         Object.keys(the_function.context).forEach(function (id) {
-            if (id !== "ignore") {
+            if (id !== 'ignore') {
                 var name = the_function.context[id];
                 if (name.function === the_function) {
-                    if (name.used === 0 && (
-                        name.role !== "function" ||
-                        name.function.arity !== "unary"
-                    )) {
-                        warn("unused_a", name);
+                    if (name.used === 0) {
+                        warn('unused_a', name);
                     } else if (!name.init) {
-                        warn("uninitialized_a", name);
+                        warn('uninitialized_a', name);
                     }
                 }
             }
@@ -4176,19 +4028,19 @@ var jslint = (function JSLint() {
 // Go through the token list, looking at usage of whitespace.
 
     function whitage() {
-        var closer = "(end)";
-        var free = false;
-        var left = global;
-        var margin = 0;
-        var nr_comments_skipped = 0;
-        var open = true;
-        var qmark = "";
-        var result;
-        var right;
+        var closer = '(end)',
+            free = false,
+            left = global,
+            margin = 0,
+            nr_comments_skipped = 0,
+            open = true,
+            qmark = '',
+            result,
+            right;
 
         function expected_at(at) {
             warn(
-                "expected_a_at_b_c",
+                'expected_a_at_b_c',
                 right,
                 artifact(right),
                 fudge + at,
@@ -4204,11 +4056,9 @@ var jslint = (function JSLint() {
         }
 
         function no_space_only() {
-            if (left.id !== "(global)" && (
-                left.line !== right.line || left.thru !== right.from
-            )) {
+            if (left.line !== right.line || left.thru !== right.from) {
                 warn(
-                    "unexpected_space_a_b",
+                    'unexpected_space_a_b',
                     right,
                     artifact(left),
                     artifact(right)
@@ -4220,7 +4070,7 @@ var jslint = (function JSLint() {
             if (left.line === right.line) {
                 if (left.thru !== right.from && nr_comments_skipped === 0) {
                     warn(
-                        "unexpected_space_a_b",
+                        'unexpected_space_a_b',
                         right,
                         artifact(left),
                         artifact(right)
@@ -4245,7 +4095,7 @@ var jslint = (function JSLint() {
         function one_space_only() {
             if (left.line !== right.line || left.thru + 1 !== right.from) {
                 warn(
-                    "expected_space_a_b",
+                    'expected_space_a_b',
                     right,
                     artifact(left),
                     artifact(right)
@@ -4257,7 +4107,7 @@ var jslint = (function JSLint() {
             if (left.line === right.line) {
                 if (left.thru + 1 !== right.from && nr_comments_skipped === 0) {
                     warn(
-                        "expected_space_a_b",
+                        'expected_space_a_b',
                         right,
                         artifact(left),
                         artifact(right)
@@ -4284,25 +4134,25 @@ var jslint = (function JSLint() {
             if (level > 0) {
                 margin -= level * 4;
             }
-            qmark = "";
+            qmark = '';
         }
 
         stack = [];
         tokens.forEach(function (the_token) {
             right = the_token;
-            if (right.id === "(comment)" || right.id === "(end)") {
+            if (right.id === '(comment)' || right.id === '(end)') {
                 nr_comments_skipped += 1;
             } else {
 
 // If left is an opener and right is not the closer, then push the previous
 // state. If the token following the opener is on the next line, then this is
-// an open form. If the tokens are on different lines, then it is a closed form.
+// an open form. If the tokens are on different lines, then it is a closed for.
 // Open form is more readable, with each item (statement, argument, parameter,
 // etc) starting on its own line. Closed form is more compact. Statement blocks
 // are always in open form.
 
                 var new_closer = opener[left.id];
-                if (typeof new_closer === "string") {
+                if (typeof new_closer === 'string') {
                     if (new_closer !== right.id) {
                         stack.push({
                             closer: closer,
@@ -4311,13 +4161,13 @@ var jslint = (function JSLint() {
                             open: open,
                             qmark: qmark
                         });
-                        qmark = "";
+                        qmark = '';
                         closer = new_closer;
                         if (left.line !== right.line) {
-                            free = closer === ")" && left.free;
+                            free = closer === ')' && left.free;
                             open = true;
                             margin += 4;
-                            if (right.role === "label") {
+                            if (right.role === 'label') {
                                 if (right.from !== 0) {
                                     expected_at(0);
                                 }
@@ -4328,9 +4178,9 @@ var jslint = (function JSLint() {
                                 at_margin(0);
                             }
                         } else {
-                            if (right.statement || right.role === "label") {
+                            if (right.statement || right.role === 'label') {
                                 warn(
-                                    "expected_line_break_a_b",
+                                    'expected_line_break_a_b',
                                     right,
                                     artifact(left),
                                     artifact(right)
@@ -4353,12 +4203,12 @@ var jslint = (function JSLint() {
                     }
                 } else {
 
-// If right is a closer, then pop the previous state.
+// If right is a closer, then pop the previous state,
 
                     if (right.id === closer) {
                         var previous = stack.pop();
                         margin = previous.margin;
-                        if (open && right.id !== ";") {
+                        if (open && right.id !== ';') {
                             at_margin(0);
                         } else {
                             no_space_only();
@@ -4379,14 +4229,14 @@ var jslint = (function JSLint() {
                         if (right.switch) {
                             unqmark();
                             at_margin(-4);
-                        } else if (right.role === "label") {
+                        } else if (right.role === 'label') {
                             if (right.from !== 0) {
                                 expected_at(0);
                             }
-                        } else if (left.id === ",") {
+                        } else if (left.id === ',') {
                             unqmark();
                             if (!open || (
-                                (free || closer === "]") &&
+                                (free || closer === ']') &&
                                 left.line === right.line
                             )) {
                                 one_space();
@@ -4397,49 +4247,42 @@ var jslint = (function JSLint() {
 // If right is a ternary operator, line it up on the margin. Use qmark to
 // deal with nested ternary operators.
 
-                        } else if (right.arity === "ternary") {
-                            if (right.id === "?") {
+                        } else if (right.arity === 'ternary') {
+                            if (right.id === '?') {
                                 margin += 4;
-                                qmark += "?";
+                                qmark += '?';
                             } else {
                                 result = qmark.match(rx_colons);
-                                qmark = result[1] + ":";
+                                qmark = result[1] + ':';
                                 margin -= 4 * result[2].length;
                             }
                             at_margin(0);
-                        } else if (
-                            right.arity === "binary" &&
-                            right.id === "(" &&
-                            free
-                        ) {
+                        } else if (right.arity === 'binary' && right.id === '(' && free) {
                             no_space();
                         } else if (
-                            left.id === "." ||
-                            left.id === "..." ||
-                            right.id === "," ||
-                            right.id === ";" ||
-                            right.id === ":" ||
-                            (right.arity === "binary" && (
-                                right.id === "(" ||
-                                right.id === "["
+                            left.id === '.' ||
+                            left.id === '...' ||
+                            right.id === ',' ||
+                            right.id === ';' ||
+                            right.id === ':' ||
+                            (right.arity === 'binary' && (
+                                right.id === '(' ||
+                                right.id === '['
                             )) ||
-                            (
-                                right.arity === "function" &&
-                                left.id !== "function"
-                            )
+                            (right.arity === 'function' && left.id !== 'function')
                         ) {
                             no_space_only();
-                        } else if (right.id === ".") {
+                        } else if (right.id === '.') {
                             if (left.line === right.line) {
                                 no_space();
                             } else {
                                 if (!rx_dot.test(qmark)) {
-                                    qmark += ".";
+                                    qmark += '.';
                                     margin += 4;
                                 }
                                 at_margin(0);
                             }
-                        } else if (left.id === ";") {
+                        } else if (left.id === ';') {
                             unqmark();
                             if (open) {
                                 at_margin(0);
@@ -4447,17 +4290,17 @@ var jslint = (function JSLint() {
                                 one_space();
                             }
                         } else if (
-                            left.arity === "ternary" ||
-                            left.id === "case" ||
-                            left.id === "catch" ||
-                            left.id === "else" ||
-                            left.id === "finally" ||
-                            left.id === "while" ||
-                            right.id === "catch" ||
-                            right.id === "else" ||
-                            right.id === "finally" ||
-                            (right.id === "while" && !right.statement) ||
-                            (left.id === ")" && right.id === "{")
+                            left.arity === 'ternary' ||
+                            left.id === 'case' ||
+                            left.id === 'catch' ||
+                            left.id === 'else' ||
+                            left.id === 'finally' ||
+                            left.id === 'while' ||
+                            right.id === 'catch' ||
+                            right.id === 'else' ||
+                            right.id === 'finally' ||
+                            (right.id === 'while' && !right.statement) ||
+                            (left.id === ')' && right.id === '{')
                         ) {
                             one_space_only();
                         } else if (right.statement === true) {
@@ -4467,9 +4310,9 @@ var jslint = (function JSLint() {
                                 one_space();
                             }
                         } else if (
-                            left.id === "var" ||
-                            left.id === "const" ||
-                            left.id === "let"
+                            left.id === 'var' ||
+                            left.id === 'const' ||
+                            left.id === 'let'
                         ) {
                             stack.push({
                                 closer: closer,
@@ -4478,10 +4321,10 @@ var jslint = (function JSLint() {
                                 open: open,
                                 qmark: qmark
                             });
-                            closer = ";";
+                            closer = ';';
                             free = false;
                             open = left.open;
-                            qmark = "";
+                            qmark = '';
                             if (open) {
                                 margin = margin + 4;
                                 at_margin(0);
@@ -4495,31 +4338,31 @@ var jslint = (function JSLint() {
                             spaceop[left.id] === true ||
                             spaceop[right.id] === true ||
                             (
-                                left.arity === "binary" &&
-                                (left.id === "+" || left.id === "-")
+                                left.arity === 'binary' &&
+                                (left.id === '+' || left.id === '-')
                             ) ||
                             (
-                                right.arity === "binary" &&
-                                (right.id === "+" || right.id === "-")
+                                right.arity === 'binary' &&
+                                (right.id === '+' || right.id === '-')
                             ) ||
-                            left.id === "function" ||
-                            left.id === ":" ||
+                            left.id === 'function' ||
+                            left.id === ':' ||
                             (
                                 (
                                     left.identifier ||
-                                    left.id === "(string)" ||
-                                    left.id === "(number)"
+                                    left.id === '(string)' ||
+                                    left.id === '(number)'
                                 ) &&
                                 (
                                     right.identifier ||
-                                    right.id === "(string)" ||
-                                    right.id === "(number)"
+                                    right.id === '(string)' ||
+                                    right.id === '(number)'
                                 )
                             ) ||
-                            (left.arity === "statement" && right.id !== ";")
+                            (left.arity === 'statement' && right.id !== ';')
                         ) {
                             one_space();
-                        } else if (left.arity === "unary" && left.id !== "`") {
+                        } else if (left.arity === 'unary') {
                             no_space_only();
                         }
                     }
@@ -4546,7 +4389,6 @@ var jslint = (function JSLint() {
             block_stack = [];
             declared_globals = empty();
             directive_mode = true;
-            directives = [];
             early_stop = true;
             export_mode = false;
             fudge = (option.fudge)
@@ -4554,7 +4396,7 @@ var jslint = (function JSLint() {
                 : 0;
             functions = [];
             global = {
-                id: "(global)",
+                id: '(global)',
                 body: true,
                 context: empty(),
                 from: 0,
@@ -4592,34 +4434,34 @@ var jslint = (function JSLint() {
             });
             tokenize(source);
             advance();
-            if (tokens[0].id === "{" || tokens[0].id === "[") {
+            if (tokens[0].id === '{' || tokens[0].id === '[') {
                 json_mode = true;
                 tree = json_value();
-                advance("(end)");
+                advance('(end)');
             } else {
 
 // Because browsers encourage combining of script files, the first token might
 // be a semicolon to defend against a missing semicolon in the preceding file.
 
                 if (option.browser) {
-                    if (next_token.id === ";") {
-                        advance(";");
+                    if (next_token.id === ';') {
+                        advance(';');
                     }
                 } else {
 
 // If we are not in a browser, then the file form of strict pragma may be used.
 
                     if (
-                        next_token.id === "(string)" &&
-                        next_token.value === "use strict"
+                        next_token.id === '(string)' &&
+                        next_token.value === 'use strict'
                     ) {
-                        advance("(string)");
-                        advance(";");
+                        advance('(string)');
+                        advance(';');
                         global.strict = true;
                     }
                 }
                 tree = statements();
-                advance("(end)");
+                advance('(end)');
                 functionage = global;
                 walk_statement(tree);
                 uninitialized_and_unused();
@@ -4627,22 +4469,14 @@ var jslint = (function JSLint() {
                     whitage();
                 }
             }
-            if (!option.browser) {
-                directives.forEach(function (comment) {
-                    if (comment.directive === "global") {
-                        warn("missing_browser", comment);
-                    }
-                });
-            }
             early_stop = false;
         } catch (e) {
-            if (e.name !== "JSLintError") {
+            if (e.name !== 'JSLintError') {
                 warnings.push(e);
             }
         }
         return {
-            directives: directives,
-            edition: "2016-05-10",
+            edition: "2015-10-29",
             functions: functions,
             global: global,
             id: "(JSLint)",
@@ -4662,5 +4496,3 @@ var jslint = (function JSLint() {
         };
     };
 }());
-
-/*node module.exports = jslint;*/
